BROWSE {httr}	R Documentation

Open specified url in browser.

Description

(This isn't really a http verb, but it seems to follow the same format).

Usage

BROWSE(url = NULL, config = list(), ..., handle = NULL)

Arguments

url
the url of the page to retrieve

config
All configuration options are ignored because the request is handled by the browser, not RCurl.

...
Further named parameters, such as query, path, etc, passed on to modify_url(). Unnamed parameters will be combined with config().

handle
The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool() based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool() for more details.

Details

Only works in interactive sessions.

Value

A response() object.

See Also

Other http methods: DELETE(), GET(), HEAD(), PATCH(), POST(), PUT(), VERB()

Examples

BROWSE("http://google.com")
BROWSE("http://had.co.nz")

[Package httr version 1.4.7 Index]