VERB {httr}	R Documentation

VERB a url.

Description

Use an arbitrary verb.

Usage

VERB(verb, url = NULL, config = list(), ..., handle = NULL)

Arguments

verb
Name of verb to use.

url
the url of the page to retrieve

config
Additional configuration settings such as http authentication (authenticate), additional headers (add_headers), cookies (set_cookies) etc. See config for full details and list of helpers.

...
Further named parameters, such as query, path, etc, passed on to modify_url. Unnamed parameters will be combined with config.

handle
The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool for more details.

See Also

Other http methods: BROWSE; DELETE; GET; HEAD

Examples

r <- VERB("PROPFIND", "http://svn.r-project.org/R/tags/",
  add_headers(depth = 1), verbose())
stop_for_status(r)
content(r)

[Package httr version 0.6.1 Index]