Add additional headers to a request.

Wikipedia provides a useful list of common http headers: http://en.wikipedia.org/wiki/List_of_HTTP_header_fields.

add_headers(..., .headers = character())

Arguments

...
named header values. To stop an existing header from being set, pass an empty string: "".

.headers
a named character vector

See also

Examples

add_headers(a = 1, b = 2)

#> <request>
#> Headers:
#> * a: 1
#> * b: 2

add_headers(.headers = c(a = "1", b = "2"))

#> <request>
#> Headers:
#> * a: 1
#> * b: 2

GET("http://httpbin.org/headers")

#> Response [http://httpbin.org/headers]
#>   Date: 2020-07-20 13:42
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 289 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.64.1 r-curl/4.3 httr/1.4.1.9000", 
#>     "X-Amzn-Trace-Id": "Root=1-5f159f2b-9d81353a82d5dfb6427ca7ec"
#>   }
#> }

# Add arbitrary headers
GET(
  "http://httpbin.org/headers",
  add_headers(version = version$version.string)
)

#> Response [http://httpbin.org/headers]
#>   Date: 2020-07-20 13:42
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 337 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.64.1 r-curl/4.3 httr/1.4.1.9000", 
#>     "Version": "R version 4.0.2 (2020-06-22)", 
#>     "X-Amzn-Trace-Id": "Root=1-5f159f2b-a402f98896c99f74d7befbde"
#>   }
#> }

# Override default headers with empty strings
GET("http://httpbin.org/headers", add_headers(Accept = ""))

#> Response [http://httpbin.org/headers]
#>   Date: 2020-07-20 13:42
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 222 B
#> {
#>   "headers": {
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.64.1 r-curl/4.3 httr/1.4.1.9000", 
#>     "X-Amzn-Trace-Id": "Root=1-5f159f2b-09f2e106e138e758ca6e4496"
#>   }
#> }