handle {httr}	R Documentation

Create a handle tied to a particular host.

Description

This handle preserves settings and cookies across multiple requests. It is the foundation of all requests performed through the httr package, although it will mostly be hidden from the user.

Usage

handle(url, cookies = TRUE)

Arguments

url
full url to site

cookies
if TRUE (the default), maintain cookies across requests.

Examples

handle("http://google.com")
handle("https://google.com")

h <- handle("http://google.com")
GET(handle = h)
# Should see cookies sent back to server
GET(handle = h, config = verbose())

h <- handle("http://google.com", cookies = FALSE)
GET(handle = h)$cookies

[Package httr version 0.6.1 Index]