http_status {httr}	R Documentation

Give information on the status of a request.

Description

Extract the http status code and convert it into a human readable message.

Usage

http_status(x)

Arguments

x
a request object or a number.

Details

http servers send a status code with the response to each request. This code gives information regarding the outcome of the execution of the request on the server. Roughly speaking, codes in the 100s and 200s mean the request was successfully executed; codes in the 300s mean the page was redirected; codes in the 400s mean there was a mistake in the way the client sent the request; codes in the 500s mean the server failed to fulfill an apparently valid request. More details on the codes can be found at ⁠http://en.wikipedia.org/wiki/Http_error_codes⁠.

Value

If the status code does not match a known status, an error. Otherwise, a list with components

category
the broad category of the status

message
the meaning of the status code

See Also

Other response methods: content(), http_error(), response(), stop_for_status()

Examples

http_status(100)
http_status(404)

## Not run: 
x <- GET("http://httpbin.org/status/200")
http_status(x)

http_status(GET("http://httpbin.org/status/300"))
http_status(GET("http://httpbin.org/status/301"))
http_status(GET("http://httpbin.org/status/404"))

# errors out on unknown status
http_status(GET("http://httpbin.org/status/320"))

## End(Not run)

[Package httr version 1.4.7 Index]