List available options.

This function lists all available options for config(). It provides both the short R name which you use with httr, and the longer Curl name, which is useful when searching the documentation. curl_doc opens a link to the libcurl documentation for an option in your browser.

httr_options(matches)

curl_docs(x)

Arguments

matches
If not missing, this restricts the output so that either the httr or curl option matches this regular expression.

x
An option name (either short or full).

Value

A data frame with three columns:

httr

The short name used in httr

libcurl

The full name used by libcurl

type

The type of R object that the option accepts

Details

RCurl and httr use slightly different names to libcurl: the initial CURLOPT_ is removed, all underscores are converted to periods and the option is given in lower case. Thus "CURLOPT_SSLENGINE_DEFAULT" becomes "sslengine.default".

Examples

httr_options()

#>                           httr                            libcurl     type
#> 1         abstract_unix_socket       CURLOPT_ABSTRACT_UNIX_SOCKET   string
#> 2              accept_encoding            CURLOPT_ACCEPT_ENCODING   string
#> 3             accepttimeout_ms           CURLOPT_ACCEPTTIMEOUT_MS  integer
#> 4                address_scope              CURLOPT_ADDRESS_SCOPE  integer
#> 5                       append                     CURLOPT_APPEND  integer
#> 6                  autoreferer                CURLOPT_AUTOREFERER  integer
#> 7                   buffersize                 CURLOPT_BUFFERSIZE  integer
#> 8                       cainfo                     CURLOPT_CAINFO   string
#> 9                       capath                     CURLOPT_CAPATH   string
#> 10                    certinfo                   CURLOPT_CERTINFO  integer
#> 11          chunk_bgn_function         CURLOPT_CHUNK_BGN_FUNCTION function
#> 12                  chunk_data                 CURLOPT_CHUNK_DATA   string
#> 13          chunk_end_function         CURLOPT_CHUNK_END_FUNCTION function
#> 14             closesocketdata            CURLOPT_CLOSESOCKETDATA   string
#> 15         closesocketfunction        CURLOPT_CLOSESOCKETFUNCTION function
#> 16                connect_only               CURLOPT_CONNECT_ONLY  integer
#> 17                  connect_to                 CURLOPT_CONNECT_TO   string
#> 18              connecttimeout             CURLOPT_CONNECTTIMEOUT  integer
#> 19           connecttimeout_ms          CURLOPT_CONNECTTIMEOUT_MS  integer
#> 20  conv_from_network_function CURLOPT_CONV_FROM_NETWORK_FUNCTION function
#> 21     conv_from_utf8_function    CURLOPT_CONV_FROM_UTF8_FUNCTION function
#> 22    conv_to_network_function   CURLOPT_CONV_TO_NETWORK_FUNCTION function
#> 23                      cookie                     CURLOPT_COOKIE   string
#> 24                  cookiefile                 CURLOPT_COOKIEFILE   string
#> 25                   cookiejar                  CURLOPT_COOKIEJAR   string
#> 26                  cookielist                 CURLOPT_COOKIELIST   string
#> 27               cookiesession              CURLOPT_COOKIESESSION  integer
#> 28              copypostfields             CURLOPT_COPYPOSTFIELDS   string
#> 29                        crlf                       CURLOPT_CRLF  integer
#> 30                     crlfile                    CURLOPT_CRLFILE   string
#> 31               customrequest              CURLOPT_CUSTOMREQUEST   string
#> 32                   debugdata                  CURLOPT_DEBUGDATA   string
#> 33               debugfunction              CURLOPT_DEBUGFUNCTION function
#> 34            default_protocol           CURLOPT_DEFAULT_PROTOCOL   string
#> 35                 dirlistonly                CURLOPT_DIRLISTONLY  integer
#> 36           dns_cache_timeout          CURLOPT_DNS_CACHE_TIMEOUT  integer
#> 37               dns_interface              CURLOPT_DNS_INTERFACE   string
#> 38               dns_local_ip4              CURLOPT_DNS_LOCAL_IP4   string
#> 39               dns_local_ip6              CURLOPT_DNS_LOCAL_IP6   string
#> 40                 dns_servers                CURLOPT_DNS_SERVERS   string
#> 41        dns_use_global_cache       CURLOPT_DNS_USE_GLOBAL_CACHE  integer
#> 42                   egdsocket                  CURLOPT_EGDSOCKET   string
#> 43                 errorbuffer                CURLOPT_ERRORBUFFER   string
#> 44       expect_100_timeout_ms      CURLOPT_EXPECT_100_TIMEOUT_MS  integer
#> 45                 failonerror                CURLOPT_FAILONERROR  integer
#> 46                    filetime                   CURLOPT_FILETIME  integer
#> 47                fnmatch_data               CURLOPT_FNMATCH_DATA   string
#> 48            fnmatch_function           CURLOPT_FNMATCH_FUNCTION function
#> 49              followlocation             CURLOPT_FOLLOWLOCATION  integer
#> 50                forbid_reuse               CURLOPT_FORBID_REUSE  integer
#> 51               fresh_connect              CURLOPT_FRESH_CONNECT  integer
#> 52                 ftp_account                CURLOPT_FTP_ACCOUNT   string
#> 53     ftp_alternative_to_user    CURLOPT_FTP_ALTERNATIVE_TO_USER   string
#> 54     ftp_create_missing_dirs    CURLOPT_FTP_CREATE_MISSING_DIRS  integer
#> 55              ftp_filemethod             CURLOPT_FTP_FILEMETHOD  integer
#> 56        ftp_response_timeout       CURLOPT_FTP_RESPONSE_TIMEOUT  integer
#> 57            ftp_skip_pasv_ip           CURLOPT_FTP_SKIP_PASV_IP  integer
#> 58                 ftp_ssl_ccc                CURLOPT_FTP_SSL_CCC  integer
#> 59                ftp_use_eprt               CURLOPT_FTP_USE_EPRT  integer
#> 60                ftp_use_epsv               CURLOPT_FTP_USE_EPSV  integer
#> 61                ftp_use_pret               CURLOPT_FTP_USE_PRET  integer
#> 62                     ftpport                    CURLOPT_FTPPORT   string
#> 63                  ftpsslauth                 CURLOPT_FTPSSLAUTH  integer
#> 64           gssapi_delegation          CURLOPT_GSSAPI_DELEGATION  integer
#> 65                      header                     CURLOPT_HEADER  integer
#> 66                  headerdata                 CURLOPT_HEADERDATA   string
#> 67              headerfunction             CURLOPT_HEADERFUNCTION function
#> 68                   headeropt                  CURLOPT_HEADEROPT  integer
#> 69       http_content_decoding      CURLOPT_HTTP_CONTENT_DECODING  integer
#> 70      http_transfer_decoding     CURLOPT_HTTP_TRANSFER_DECODING  integer
#> 71                http_version               CURLOPT_HTTP_VERSION  integer
#> 72              http200aliases             CURLOPT_HTTP200ALIASES   string
#> 73                    httpauth                   CURLOPT_HTTPAUTH  integer
#> 74                     httpget                    CURLOPT_HTTPGET  integer
#> 75                  httpheader                 CURLOPT_HTTPHEADER   string
#> 76                    httppost                   CURLOPT_HTTPPOST   string
#> 77             httpproxytunnel            CURLOPT_HTTPPROXYTUNNEL  integer
#> 78       ignore_content_length      CURLOPT_IGNORE_CONTENT_LENGTH  integer
#> 79                  infilesize                 CURLOPT_INFILESIZE  integer
#> 80            infilesize_large           CURLOPT_INFILESIZE_LARGE   number
#> 81                   interface                  CURLOPT_INTERFACE   string
#> 82              interleavedata             CURLOPT_INTERLEAVEDATA   string
#> 83          interleavefunction         CURLOPT_INTERLEAVEFUNCTION function
#> 84                   ioctldata                  CURLOPT_IOCTLDATA   string
#> 85               ioctlfunction              CURLOPT_IOCTLFUNCTION function
#> 86                   ipresolve                  CURLOPT_IPRESOLVE  integer
#> 87                  issuercert                 CURLOPT_ISSUERCERT   string
#> 88       keep_sending_on_error      CURLOPT_KEEP_SENDING_ON_ERROR  integer
#> 89                   keypasswd                  CURLOPT_KEYPASSWD   string
#> 90                    krblevel                   CURLOPT_KRBLEVEL   string
#> 91                   localport                  CURLOPT_LOCALPORT  integer
#> 92              localportrange             CURLOPT_LOCALPORTRANGE  integer
#> 93               login_options              CURLOPT_LOGIN_OPTIONS   string
#> 94             low_speed_limit            CURLOPT_LOW_SPEED_LIMIT  integer
#> 95              low_speed_time             CURLOPT_LOW_SPEED_TIME  integer
#> 96                   mail_auth                  CURLOPT_MAIL_AUTH   string
#> 97                   mail_from                  CURLOPT_MAIL_FROM   string
#> 98                   mail_rcpt                  CURLOPT_MAIL_RCPT   string
#> 99        max_recv_speed_large       CURLOPT_MAX_RECV_SPEED_LARGE   number
#> 100       max_send_speed_large       CURLOPT_MAX_SEND_SPEED_LARGE   number
#> 101                maxconnects                CURLOPT_MAXCONNECTS  integer
#> 102                maxfilesize                CURLOPT_MAXFILESIZE  integer
#> 103          maxfilesize_large          CURLOPT_MAXFILESIZE_LARGE   number
#> 104                  maxredirs                  CURLOPT_MAXREDIRS  integer
#> 105                      netrc                      CURLOPT_NETRC  integer
#> 106                 netrc_file                 CURLOPT_NETRC_FILE   string
#> 107        new_directory_perms        CURLOPT_NEW_DIRECTORY_PERMS  integer
#> 108             new_file_perms             CURLOPT_NEW_FILE_PERMS  integer
#> 109                     nobody                     CURLOPT_NOBODY  integer
#> 110                 noprogress                 CURLOPT_NOPROGRESS  integer
#> 111                    noproxy                    CURLOPT_NOPROXY   string
#> 112                   nosignal                   CURLOPT_NOSIGNAL  integer
#> 113                 obsolete40                 CURLOPT_OBSOLETE40   string
#> 114                 obsolete72                 CURLOPT_OBSOLETE72  integer
#> 115             opensocketdata             CURLOPT_OPENSOCKETDATA   string
#> 116         opensocketfunction         CURLOPT_OPENSOCKETFUNCTION function
#> 117                   password                   CURLOPT_PASSWORD   string
#> 118                 path_as_is                 CURLOPT_PATH_AS_IS  integer
#> 119            pinnedpublickey            CURLOPT_PINNEDPUBLICKEY   string
#> 120                   pipewait                   CURLOPT_PIPEWAIT  integer
#> 121                       port                       CURLOPT_PORT  integer
#> 122                       post                       CURLOPT_POST  integer
#> 123                 postfields                 CURLOPT_POSTFIELDS   string
#> 124              postfieldsize              CURLOPT_POSTFIELDSIZE  integer
#> 125        postfieldsize_large        CURLOPT_POSTFIELDSIZE_LARGE   number
#> 126                  postquote                  CURLOPT_POSTQUOTE   string
#> 127                  postredir                  CURLOPT_POSTREDIR  integer
#> 128                  pre_proxy                  CURLOPT_PRE_PROXY   string
#> 129                   prequote                   CURLOPT_PREQUOTE   string
#> 130                    private                    CURLOPT_PRIVATE   string
#> 131               progressdata               CURLOPT_PROGRESSDATA   string
#> 132           progressfunction           CURLOPT_PROGRESSFUNCTION function
#> 133                  protocols                  CURLOPT_PROTOCOLS  integer
#> 134                      proxy                      CURLOPT_PROXY   string
#> 135               proxy_cainfo               CURLOPT_PROXY_CAINFO   string
#> 136               proxy_capath               CURLOPT_PROXY_CAPATH   string
#> 137              proxy_crlfile              CURLOPT_PROXY_CRLFILE   string
#> 138            proxy_keypasswd            CURLOPT_PROXY_KEYPASSWD   string
#> 139      proxy_pinnedpublickey      CURLOPT_PROXY_PINNEDPUBLICKEY   string
#> 140         proxy_service_name         CURLOPT_PROXY_SERVICE_NAME   string
#> 141      proxy_ssl_cipher_list      CURLOPT_PROXY_SSL_CIPHER_LIST   string
#> 142          proxy_ssl_options          CURLOPT_PROXY_SSL_OPTIONS  integer
#> 143       proxy_ssl_verifyhost       CURLOPT_PROXY_SSL_VERIFYHOST  integer
#> 144       proxy_ssl_verifypeer       CURLOPT_PROXY_SSL_VERIFYPEER  integer
#> 145              proxy_sslcert              CURLOPT_PROXY_SSLCERT   string
#> 146          proxy_sslcerttype          CURLOPT_PROXY_SSLCERTTYPE   string
#> 147               proxy_sslkey               CURLOPT_PROXY_SSLKEY   string
#> 148           proxy_sslkeytype           CURLOPT_PROXY_SSLKEYTYPE   string
#> 149           proxy_sslversion           CURLOPT_PROXY_SSLVERSION  integer
#> 150     proxy_tlsauth_password     CURLOPT_PROXY_TLSAUTH_PASSWORD   string
#> 151         proxy_tlsauth_type         CURLOPT_PROXY_TLSAUTH_TYPE   string
#> 152     proxy_tlsauth_username     CURLOPT_PROXY_TLSAUTH_USERNAME   string
#> 153        proxy_transfer_mode        CURLOPT_PROXY_TRANSFER_MODE  integer
#> 154                  proxyauth                  CURLOPT_PROXYAUTH  integer
#> 155                proxyheader                CURLOPT_PROXYHEADER   string
#> 156              proxypassword              CURLOPT_PROXYPASSWORD   string
#> 157                  proxyport                  CURLOPT_PROXYPORT  integer
#> 158                  proxytype                  CURLOPT_PROXYTYPE  integer
#> 159              proxyusername              CURLOPT_PROXYUSERNAME   string
#> 160               proxyuserpwd               CURLOPT_PROXYUSERPWD   string
#> 161                        put                        CURLOPT_PUT  integer
#> 162                      quote                      CURLOPT_QUOTE   string
#> 163                random_file                CURLOPT_RANDOM_FILE   string
#> 164                      range                      CURLOPT_RANGE   string
#> 165                   readdata                   CURLOPT_READDATA   string
#> 166               readfunction               CURLOPT_READFUNCTION function
#> 167            redir_protocols            CURLOPT_REDIR_PROTOCOLS  integer
#> 168                    referer                    CURLOPT_REFERER   string
#> 169                    resolve                    CURLOPT_RESOLVE   string
#> 170                resume_from                CURLOPT_RESUME_FROM  integer
#> 171          resume_from_large          CURLOPT_RESUME_FROM_LARGE   number
#> 172           rtsp_client_cseq           CURLOPT_RTSP_CLIENT_CSEQ  integer
#> 173               rtsp_request               CURLOPT_RTSP_REQUEST  integer
#> 174           rtsp_server_cseq           CURLOPT_RTSP_SERVER_CSEQ  integer
#> 175            rtsp_session_id            CURLOPT_RTSP_SESSION_ID   string
#> 176            rtsp_stream_uri            CURLOPT_RTSP_STREAM_URI   string
#> 177             rtsp_transport             CURLOPT_RTSP_TRANSPORT   string
#> 178                    sasl_ir                    CURLOPT_SASL_IR  integer
#> 179                   seekdata                   CURLOPT_SEEKDATA   string
#> 180               seekfunction               CURLOPT_SEEKFUNCTION function
#> 181               service_name               CURLOPT_SERVICE_NAME   string
#> 182                      share                      CURLOPT_SHARE   string
#> 183                sockoptdata                CURLOPT_SOCKOPTDATA   string
#> 184            sockoptfunction            CURLOPT_SOCKOPTFUNCTION function
#> 185          socks5_gssapi_nec          CURLOPT_SOCKS5_GSSAPI_NEC  integer
#> 186      socks5_gssapi_service      CURLOPT_SOCKS5_GSSAPI_SERVICE   string
#> 187             ssh_auth_types             CURLOPT_SSH_AUTH_TYPES  integer
#> 188    ssh_host_public_key_md5    CURLOPT_SSH_HOST_PUBLIC_KEY_MD5   string
#> 189                ssh_keydata                CURLOPT_SSH_KEYDATA   string
#> 190            ssh_keyfunction            CURLOPT_SSH_KEYFUNCTION function
#> 191             ssh_knownhosts             CURLOPT_SSH_KNOWNHOSTS   string
#> 192        ssh_private_keyfile        CURLOPT_SSH_PRIVATE_KEYFILE   string
#> 193         ssh_public_keyfile         CURLOPT_SSH_PUBLIC_KEYFILE   string
#> 194            ssl_cipher_list            CURLOPT_SSL_CIPHER_LIST   string
#> 195               ssl_ctx_data               CURLOPT_SSL_CTX_DATA   string
#> 196           ssl_ctx_function           CURLOPT_SSL_CTX_FUNCTION function
#> 197            ssl_enable_alpn            CURLOPT_SSL_ENABLE_ALPN  integer
#> 198             ssl_enable_npn             CURLOPT_SSL_ENABLE_NPN  integer
#> 199             ssl_falsestart             CURLOPT_SSL_FALSESTART  integer
#> 200                ssl_options                CURLOPT_SSL_OPTIONS  integer
#> 201        ssl_sessionid_cache        CURLOPT_SSL_SESSIONID_CACHE  integer
#> 202             ssl_verifyhost             CURLOPT_SSL_VERIFYHOST  integer
#> 203             ssl_verifypeer             CURLOPT_SSL_VERIFYPEER  integer
#> 204           ssl_verifystatus           CURLOPT_SSL_VERIFYSTATUS  integer
#> 205                    sslcert                    CURLOPT_SSLCERT   string
#> 206                sslcerttype                CURLOPT_SSLCERTTYPE   string
#> 207                  sslengine                  CURLOPT_SSLENGINE   string
#> 208          sslengine_default          CURLOPT_SSLENGINE_DEFAULT  integer
#> 209                     sslkey                     CURLOPT_SSLKEY   string
#> 210                 sslkeytype                 CURLOPT_SSLKEYTYPE   string
#> 211                 sslversion                 CURLOPT_SSLVERSION  integer
#> 212                     stderr                     CURLOPT_STDERR   string
#> 213             stream_depends             CURLOPT_STREAM_DEPENDS   string
#> 214           stream_depends_e           CURLOPT_STREAM_DEPENDS_E   string
#> 215              stream_weight              CURLOPT_STREAM_WEIGHT  integer
#> 216   suppress_connect_headers   CURLOPT_SUPPRESS_CONNECT_HEADERS  integer
#> 217               tcp_fastopen               CURLOPT_TCP_FASTOPEN  integer
#> 218              tcp_keepalive              CURLOPT_TCP_KEEPALIVE  integer
#> 219               tcp_keepidle               CURLOPT_TCP_KEEPIDLE  integer
#> 220              tcp_keepintvl              CURLOPT_TCP_KEEPINTVL  integer
#> 221                tcp_nodelay                CURLOPT_TCP_NODELAY  integer
#> 222              telnetoptions              CURLOPT_TELNETOPTIONS   string
#> 223               tftp_blksize               CURLOPT_TFTP_BLKSIZE  integer
#> 224            tftp_no_options            CURLOPT_TFTP_NO_OPTIONS  integer
#> 225              timecondition              CURLOPT_TIMECONDITION  integer
#> 226                    timeout                    CURLOPT_TIMEOUT  integer
#> 227                 timeout_ms                 CURLOPT_TIMEOUT_MS  integer
#> 228                  timevalue                  CURLOPT_TIMEVALUE  integer
#> 229           tlsauth_password           CURLOPT_TLSAUTH_PASSWORD   string
#> 230               tlsauth_type               CURLOPT_TLSAUTH_TYPE   string
#> 231           tlsauth_username           CURLOPT_TLSAUTH_USERNAME   string
#> 232          transfer_encoding          CURLOPT_TRANSFER_ENCODING  integer
#> 233               transfertext               CURLOPT_TRANSFERTEXT  integer
#> 234           unix_socket_path           CURLOPT_UNIX_SOCKET_PATH   string
#> 235          unrestricted_auth          CURLOPT_UNRESTRICTED_AUTH  integer
#> 236                     upload                     CURLOPT_UPLOAD  integer
#> 237                        url                        CURLOPT_URL   string
#> 238                    use_ssl                    CURLOPT_USE_SSL  integer
#> 239                  useragent                  CURLOPT_USERAGENT   string
#> 240                   username                   CURLOPT_USERNAME   string
#> 241                    userpwd                    CURLOPT_USERPWD   string
#> 242                    verbose                    CURLOPT_VERBOSE  integer
#> 243              wildcardmatch              CURLOPT_WILDCARDMATCH  integer
#> 244                  writedata                  CURLOPT_WRITEDATA   string
#> 245              writefunction              CURLOPT_WRITEFUNCTION function
#> 246           xferinfofunction           CURLOPT_XFERINFOFUNCTION function
#> 247             xoauth2_bearer             CURLOPT_XOAUTH2_BEARER   string

httr_options("post")

#>                    httr                     libcurl    type
#> 28       copypostfields      CURLOPT_COPYPOSTFIELDS  string
#> 76             httppost            CURLOPT_HTTPPOST  string
#> 122                post                CURLOPT_POST integer
#> 123          postfields          CURLOPT_POSTFIELDS  string
#> 124       postfieldsize       CURLOPT_POSTFIELDSIZE integer
#> 125 postfieldsize_large CURLOPT_POSTFIELDSIZE_LARGE  number
#> 126           postquote           CURLOPT_POSTQUOTE  string
#> 127           postredir           CURLOPT_POSTREDIR integer

# Use curl_docs to read the curl documentation for each option.
# You can use either the httr or curl option name.
curl_docs("userpwd")

#> Please point your browser to the following url: 

#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html

curl_docs("CURLOPT_USERPWD")

#> Please point your browser to the following url: 

#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html