oauth_app {httr}	R Documentation

Create an OAuth application.

Description

The OAuth framework doesn't match perfectly to use from R. Each user of the package for a particular OAuth enabled site must create their own application. See the demos for instructions on how to do this for linkedin, twitter, vimeo, facebook, github and google.

Usage

oauth_app(appname, key, secret = NULL)

Arguments

appname
name of the application. This is not used for OAuth, but is used to make it easier to identify different applications and provide a consistent way of storing secrets in environment variables.

key
consumer key (equivalent to a user name)

secret
consumer secret. This is equivalent to a password and should not be stored in publicly visible code. As a convenient shortcut, if secret is missing, we'll look in the environment variable APPNAME_CONSUMER_SECRET. Use NULL to not store a secret: this is useful if you're relying on cached OAuth tokens.

See Also

Other OAuth: oauth1.0_token; oauth2.0_token; oauth_endpoint; oauth_service_token

Examples

## Not run: 
# These work on my computer because I have the right envvars set up
linkedin_app <- oauth_app("linkedin", key = "outmkw3859gy")
github_app <- oauth_app("github", "56b637a5baffac62cad9")

## End(Not run)

# If you're relying on caching, supply an explicit NULL to
# suppress the warning message
oauth_app("my_app", "mykey")
oauth_app("my_app", "mykey", NULL)

[Package httr version 0.6.1 Index]