oauth_exchanger {httr}	R Documentation

Walk the user through the OAuth2 dance without a local webserver.

Description

This performs a similar function to oauth_listener(), but without running a local webserver. This manual process can be useful in situations where the user is remotely accessing the machine outside a browser (say via ssh) or when it's not possible to successfully receive a callback (such as when behind a firewall).

Usage

oauth_exchanger(request_url)

Arguments

request_url
the url to provide to the user

Details

This function should generally not be called directly by the user.

[Package httr version 1.4.7 Index]