Get url HEADers.

Get url HEADers.

HEAD(url = NULL, config = list(), ..., handle = NULL)

Arguments

url
the url of the page to retrieve

config
Additional configuration settings such as http authentication (authenticate()), additional headers (add_headers()), cookies (set_cookies()) etc. See config() for full details and list of helpers.

...
Further named parameters, such as query, path, etc, passed on to modify_url(). Unnamed parameters will be combined with config().

handle
The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool() based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool() for more details.

Value

A response() object.

RFC2616

The HEAD method is identical to GET except that the server MUST NOT return a message-body in the response. The metainformation contained in the HTTP headers in response to a HEAD request SHOULD be identical to the information sent in response to a GET request. This method can be used for obtaining metainformation about the entity implied by the request without transferring the entity-body itself. This method is often used for testing hypertext links for validity, accessibility, and recent modification.

The response to a HEAD request MAY be cacheable in the sense that the information contained in the response MAY be used to update a previously cached entity from that resource. If the new field values indicate that the cached entity differs from the current entity (as would be indicated by a change in Content-Length, Content-MD5, ETag or Last-Modified), then the cache MUST treat the cache entry as stale.

See also

Examples

HEAD("http://google.com")

#> Response [http://www.google.com/]
#>   Date: 2020-07-20 13:22
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>

headers(HEAD("http://google.com"))

#> $`content-type`
#> [1] "text/html; charset=ISO-8859-1"
#> 
#> $date
#> [1] "Mon, 20 Jul 2020 13:22:23 GMT"
#> 
#> $server
#> [1] "gws"
#> 
#> $`x-xss-protection`
#> [1] "0"
#> 
#> $`x-frame-options`
#> [1] "SAMEORIGIN"
#> 
#> $`transfer-encoding`
#> [1] "chunked"
#> 
#> $expires
#> [1] "Mon, 20 Jul 2020 13:22:23 GMT"
#> 
#> $`cache-control`
#> [1] "private"
#> 
#> $`set-cookie`
#> [1] "1P_JAR=2020-07-20-13; expires=Wed, 19-Aug-2020 13:22:23 GMT; path=/; domain=.google.com; Secure"
#> 
#> attr(,"class")
#> [1] "insensitive" "list"