Get url HEADers.

Get url HEADers.

HEAD(url = NULL, config = list(), ..., handle = NULL)

Arguments

url
the url of the page to retrieve

config
Additional configuration settings such as http authentication (authenticate), additional headers (add_headers), cookies (set_cookies) etc. See config for full details and list of helpers.

...
Further named parameters, such as query, path, etc, passed on to modify_url. Unnamed parameters will be combined with config.

handle
The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool for more details.

RFC2616

The HEAD method is identical to GET except that the server MUST NOT return a message-body in the response. The metainformation contained in the HTTP headers in response to a HEAD request SHOULD be identical to the information sent in response to a GET request. This method can be used for obtaining metainformation about the entity implied by the request without transferring the entity-body itself. This method is often used for testing hypertext links for validity, accessibility, and recent modification.

The response to a HEAD request MAY be cacheable in the sense that the information contained in the response MAY be used to update a previously cached entity from that resource. If the new field values indicate that the cached entity differs from the current entity (as would be indicated by a change in Content-Length, Content-MD5, ETag or Last-Modified), then the cache MUST treat the cache entry as stale.

See also

Other http methods: BROWSE, DELETE, GET, PATCH, POST, PUT, VERB

Examples

HEAD("http://google.com")

#> Response [http://www.google.com/]
#>   Date: 2017-08-01 19:59
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>

headers(HEAD("http://google.com"))

#> $date
#> [1] "Tue, 01 Aug 2017 19:59:41 GMT"
#> 
#> $expires
#> [1] "-1"
#> 
#> $`cache-control`
#> [1] "private, max-age=0"
#> 
#> $`content-type`
#> [1] "text/html; charset=ISO-8859-1"
#> 
#> $`content-encoding`
#> [1] "gzip"
#> 
#> $server
#> [1] "gws"
#> 
#> $`content-length`
#> [1] "4953"
#> 
#> $`x-xss-protection`
#> [1] "1; mode=block"
#> 
#> $`x-frame-options`
#> [1] "SAMEORIGIN"
#> 
#> attr(,"class")
#> [1] "insensitive" "list"