Add additional headers to a request.

Wikipedia provides a useful list of common http headers: http://en.wikipedia.org/wiki/List_of_HTTP_header_fields.

add_headers(..., .headers = character())

Arguments

...
named header values. To stop an existing header from being set, pass an empty string: "".

.headers
a named character vector

See also

accept and content_type for convenience functions for setting accept and content-type headers.

Other config: authenticate, config, set_cookies, timeout, use_proxy, user_agent, verbose

Examples

add_headers(a = 1, b = 2)

#> <request>
#> Headers:
#> * a: 1
#> * b: 2

add_headers(.headers = c(a = "1", b = "2"))

#> <request>
#> Headers:
#> * a: 1
#> * b: 2

GET("http://httpbin.org/headers")

#> Response [http://httpbin.org/headers]
#>   Date: 2017-08-01 19:59
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 251 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "gzip, deflate", 
#>     "Connection": "close", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.54.0 r-curl/2.8.1 httr/1.2.1.9000"
#>   }
#> }

# Add arbitrary headers
GET("http://httpbin.org/headers",
 add_headers(version = version$version.string))

#> Response [http://httpbin.org/headers]
#>   Date: 2017-08-01 19:59
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 299 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "gzip, deflate", 
#>     "Connection": "close", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.54.0 r-curl/2.8.1 httr/1.2.1.9000", 
#>     "Version": "R version 3.4.1 (2017-06-30)"
#>   }
#> }

# Override default headers with empty strings
GET("http://httpbin.org/headers", add_headers(Accept = ""))

#> Response [http://httpbin.org/headers]
#>   Date: 2017-08-01 19:59
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 184 B
#> {
#>   "headers": {
#>     "Accept-Encoding": "gzip, deflate", 
#>     "Connection": "close", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.54.0 r-curl/2.8.1 httr/1.2.1.9000"
#>   }
#> }