Add additional headers to a request.

Wikipedia provides a useful list of common http headers: https://en.wikipedia.org/wiki/List_of_HTTP_header_fields.

Usage

add_headers(..., .headers = character())

Arguments

...

named header values. To stop an existing header from being set, pass an empty string: "".

.headers

a named character vector

See also

Examples

add_headers(a = 1, b = 2)
#> <request>
#> Headers:
#> * a: 1
#> * b: 2
add_headers(.headers = c(a = "1", b = "2"))
#> <request>
#> Headers:
#> * a: 1
#> * b: 2

GET("http://httpbin.org/headers")
#> Response [http://httpbin.org/headers]
#>   Date: 2023-02-22 18:21
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 301 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip, br, zstd", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.81.0 r-curl/5.0.0 httr/1.4.4.9000", 
#>     "X-Amzn-Trace-Id": "Root=1-63f65d12-0a2062e05a8b35990e199cc7"
#>   }
#> }

# Add arbitrary headers
GET(
  "http://httpbin.org/headers",
  add_headers(version = version$version.string)
)
#> Response [http://httpbin.org/headers]
#>   Date: 2023-02-22 18:21
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 349 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip, br, zstd", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.81.0 r-curl/5.0.0 httr/1.4.4.9000", 
#>     "Version": "R version 4.2.2 (2022-10-31)", 
#>     "X-Amzn-Trace-Id": "Root=1-63f65d12-5084026052f219cd4219594a"
#>   }
#> }

# Override default headers with empty strings
GET("http://httpbin.org/headers", add_headers(Accept = ""))
#> Response [http://httpbin.org/headers]
#>   Date: 2023-02-22 18:21
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 234 B
#> {
#>   "headers": {
#>     "Accept-Encoding": "deflate, gzip, br, zstd", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.81.0 r-curl/5.0.0 httr/1.4.4.9000", 
#>     "X-Amzn-Trace-Id": "Root=1-63f65d12-3f7f35cf71e9b96471eb1d40"
#>   }
#> }