authenticate {httr}	R Documentation

Use http authentication.

Description

It's not obvious how to turn authentication off after using it, so I recommend using custom handles with authentication.

Usage

authenticate(user, password, type = "basic")

Arguments

user
user name

password
password

type
type of HTTP authentication. Should be one of the following types supported by Curl: basic, digest, digest_ie, gssnegotiate, ntlm, any. It defaults to "basic", the most common type.

See Also

Other config: add_headers(), config(), set_cookies(), timeout(), use_proxy(), user_agent(), verbose()

Examples

## Not run: 
GET("http://httpbin.org/basic-auth/user/passwd")
GET(
  "http://httpbin.org/basic-auth/user/passwd",
  authenticate("user", "passwd")
)

## End(Not run)

[Package httr version 1.4.7 Index]