Set curl options.

Generally you should only need to use this function to set CURL options directly if there isn't already a helpful wrapper function, like set_cookies(), add_headers() or authenticate(). To use this function effectively requires some knowledge of CURL, and CURL options. Use httr_options() to see a complete list of available options. To see the libcurl documentation for a given option, use curl_docs().

Usage

config(..., token = NULL)

Arguments

...

named Curl options.

token

An OAuth token (1.0 or 2.0)

Details

Unlike Curl (and RCurl), all configuration options are per request, not per handle.

See also

Examples

# There are a number of ways to modify the configuration of a request
# * you can add directly to a request
HEAD("https://www.google.com", verbose())
#> Response [https://www.google.com/]
#>   Date: 2026-08-31 23:02
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>

# * you can wrap with with_config()
with_config(verbose(), HEAD("https://www.google.com"))
#> Response [https://www.google.com/]
#>   Date: 2026-08-31 23:02
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>

# * you can set global with set_config()
old <- set_config(verbose())
HEAD("https://www.google.com")
#> Response [https://www.google.com/]
#>   Date: 2026-08-31 23:02
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>
# and re-establish the previous settings with
set_config(old, override = TRUE)
HEAD("https://www.google.com")
#> Response [https://www.google.com/]
#>   Date: 2026-08-31 23:02
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>
# or
reset_config()
HEAD("https://www.google.com")
#> Response [https://www.google.com/]
#>   Date: 2026-08-31 23:02
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>

# If available, you should use a friendly httr wrapper over RCurl
# options. But you can pass Curl options (as listed in httr_options())
# in config
HEAD("https://www.google.com/", config(verbose = TRUE))
#> Response [https://www.google.com/]
#>   Date: 2026-08-31 23:02
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#> <EMPTY BODY>