Set content-type and accept headers.

These are convenient wrappers aroud add_headers().

content_type(type)

content_type_json()

content_type_xml()

accept(type)

accept_json()

accept_xml()

Arguments

type
A mime type or a file extension. If a file extension (i.e. starts with .) will guess the mime type using mime::guess_type().

Details

accept_json/accept_xml and content_type_json/content_type_xml are useful shortcuts to ask for json or xml responses or tell the server you are sending json/xml.

Examples

GET("http://httpbin.org/headers")

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 221 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }

GET("http://httpbin.org/headers", accept_json())

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 189 B
#> {
#>   "headers": {
#>     "Accept": "application/json", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }

GET("http://httpbin.org/headers", accept("text/csv"))

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 181 B
#> {
#>   "headers": {
#>     "Accept": "text/csv", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }

GET("http://httpbin.org/headers", accept(".doc"))

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 191 B
#> {
#>   "headers": {
#>     "Accept": "application/msword", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }

GET("http://httpbin.org/headers", content_type_xml())

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 261 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Content-Type": "application/xml", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }

GET("http://httpbin.org/headers", content_type("text/csv"))

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 254 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Content-Type": "text/csv", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }

GET("http://httpbin.org/headers", content_type(".xml"))

#> Response [http://httpbin.org/headers]
#>   Date: 2019-07-27 13:05
#>   Status: 200
#>   Content-Type: application/json
#>   Size: 261 B
#> {
#>   "headers": {
#>     "Accept": "application/json, text/xml, application/xml, */*", 
#>     "Accept-Encoding": "deflate, gzip", 
#>     "Content-Type": "application/xml", 
#>     "Host": "httpbin.org", 
#>     "User-Agent": "libcurl/7.43.0 r-curl/4.0 httr/1.4.0.9000"
#>   }
#> }