Create a handle tied to a particular host.

This handle preserves settings and cookies across multiple requests. It is the foundation of all requests performed through the httr package, although it will mostly be hidden from the user.

handle(url, cookies = TRUE)

Arguments

url
full url to site

cookies
DEPRECATED

Note

Because of the way argument dispatch works in R, using handle() in the http methods (See GET()) will cause problems when trying to pass configuration arguments (See examples below). Directly specifying the handle when using http methods is not recommended in general, since the selection of the correct handle is taken care of when the user passes an url (See handle_pool()).

Examples

handle("http://google.com")

#> Host: http://google.com <NA>

handle("https://google.com")

#> Host: https://google.com <NA>

h <- handle("http://google.com")
GET(handle = h)

#> Response [http://www.google.com/]
#>   Date: 2021-02-08 15:25
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#>   Size: 13.6 kB
#> <!doctype html><html itemscope="" itemtype="http://schema.org/WebPage" lang="...
#> google.lc=[];google.li=0;google.getEI=function(a){for(var b;a&&(!a.getAttribu...
#> document.documentElement.addEventListener("submit",function(b){var a;if(a=b.t...
#> var a=window.location,b=a.href.indexOf("#");if(0<=b){var c=a.href.substring(b...
#> </style><style>body,td,a,p,.h{font-family:arial,sans-serif}body{margin:0;over...
#> if (!iesg){document.f&&document.f.q.focus();document.gbqf&&document.gbqf.q.fo...
#> }
#> })();</script><div id="mngb"><div id=gbar><nobr><b class=gb1>Search</b> <a cl...
#> else top.location='/doodles/';};})();</script><input value="AINFCbYAAAAAYCFl3...
#> var d=this||self,e=/^[\w+/_-]+[=]{0,2}$/,f=null,g=function(a){return(a=a.quer...
#> ...

# Should see cookies sent back to server
GET(handle = h, config = verbose())

#> Response [http://www.google.com/]
#>   Date: 2021-02-08 15:25
#>   Status: 200
#>   Content-Type: text/html; charset=ISO-8859-1
#>   Size: 13.6 kB
#> <!doctype html><html itemscope="" itemtype="http://schema.org/WebPage" lang="...
#> google.lc=[];google.li=0;google.getEI=function(a){for(var b;a&&(!a.getAttribu...
#> document.documentElement.addEventListener("submit",function(b){var a;if(a=b.t...
#> var a=window.location,b=a.href.indexOf("#");if(0<=b){var c=a.href.substring(b...
#> </style><style>body,td,a,p,.h{font-family:arial,sans-serif}body{margin:0;over...
#> if (!iesg){document.f&&document.f.q.focus();document.gbqf&&document.gbqf.q.fo...
#> }
#> })();</script><div id="mngb"><div id=gbar><nobr><b class=gb1>Search</b> <a cl...
#> else top.location='/doodles/';};})();</script><input value="AINFCbYAAAAAYCFl4...
#> var d=this||self,e=/^[\w+/_-]+[=]{0,2}$/,f=null,g=function(a){return(a=a.quer...
#> ...

h <- handle("http://google.com", cookies = FALSE)

#> Warning: Cookies argument is deprecated

GET(handle = h)$cookies

#>                  domain flag path secure          expiration name
#> 1 #HttpOnly_.google.com TRUE    /  FALSE 2021-08-10 15:25:04  NID
#>                                                                                                                                                                             value
#> 1 208=LWqgSN6dwMpYBnJMQemKLL2JsgXmDgMTjzs3POqyfOeK04u7MO9a71KppYKNICvO7HmFKAQ6t8YE0JZsKmYBns4LqCf31ZtbkaRMjBAm-_xop6vecMt6_-hLNkjGKAty-Iw9-6jvQZt-nanoxcxKSoSSlj3BMgEepau_uwWwvws

if (FALSE) {
# Using the preferred way of configuring the http methods
# will not work when using handle():
GET(handle = h, timeout(10))
# Passing named arguments will work properly:
GET(handle = h, config = list(timeout(10), add_headers(Accept = "")))
}