List available options.

This function lists all available options for config(). It provides both the short R name which you use with httr, and the longer Curl name, which is useful when searching the documentation. curl_doc opens a link to the libcurl documentation for an option in your browser.

httr_options(matches)

curl_docs(x)

Arguments

matches
If not missing, this restricts the output so that either the httr or curl option matches this regular expression.

x
An option name (either short or full).

Value

A data frame with three columns:

httr

The short name used in httr

libcurl

The full name used by libcurl

type

The type of R object that the option accepts

Details

RCurl and httr use slightly different names to libcurl: the initial CURLOPT_ is removed, all underscores are converted to periods and the option is given in lower case. Thus "CURLOPT_SSLENGINE_DEFAULT" becomes "sslengine.default".

Examples

httr_options()

#>                           httr                            libcurl     type
#> 1              accept_encoding            CURLOPT_ACCEPT_ENCODING   string
#> 2             accepttimeout_ms           CURLOPT_ACCEPTTIMEOUT_MS  integer
#> 3                address_scope              CURLOPT_ADDRESS_SCOPE  integer
#> 4                       append                     CURLOPT_APPEND  integer
#> 5                  autoreferer                CURLOPT_AUTOREFERER  integer
#> 6                   buffersize                 CURLOPT_BUFFERSIZE  integer
#> 7                       cainfo                     CURLOPT_CAINFO   string
#> 8                       capath                     CURLOPT_CAPATH   string
#> 9                     certinfo                   CURLOPT_CERTINFO  integer
#> 10          chunk_bgn_function         CURLOPT_CHUNK_BGN_FUNCTION function
#> 11                  chunk_data                 CURLOPT_CHUNK_DATA   string
#> 12          chunk_end_function         CURLOPT_CHUNK_END_FUNCTION function
#> 13             closesocketdata            CURLOPT_CLOSESOCKETDATA   string
#> 14         closesocketfunction        CURLOPT_CLOSESOCKETFUNCTION function
#> 15                connect_only               CURLOPT_CONNECT_ONLY  integer
#> 16              connecttimeout             CURLOPT_CONNECTTIMEOUT  integer
#> 17           connecttimeout_ms          CURLOPT_CONNECTTIMEOUT_MS  integer
#> 18  conv_from_network_function CURLOPT_CONV_FROM_NETWORK_FUNCTION function
#> 19     conv_from_utf8_function    CURLOPT_CONV_FROM_UTF8_FUNCTION function
#> 20    conv_to_network_function   CURLOPT_CONV_TO_NETWORK_FUNCTION function
#> 21                      cookie                     CURLOPT_COOKIE   string
#> 22                  cookiefile                 CURLOPT_COOKIEFILE   string
#> 23                   cookiejar                  CURLOPT_COOKIEJAR   string
#> 24                  cookielist                 CURLOPT_COOKIELIST   string
#> 25               cookiesession              CURLOPT_COOKIESESSION  integer
#> 26              copypostfields             CURLOPT_COPYPOSTFIELDS   string
#> 27                        crlf                       CURLOPT_CRLF  integer
#> 28                     crlfile                    CURLOPT_CRLFILE   string
#> 29               customrequest              CURLOPT_CUSTOMREQUEST   string
#> 30                   debugdata                  CURLOPT_DEBUGDATA   string
#> 31               debugfunction              CURLOPT_DEBUGFUNCTION function
#> 32                 dirlistonly                CURLOPT_DIRLISTONLY  integer
#> 33           dns_cache_timeout          CURLOPT_DNS_CACHE_TIMEOUT  integer
#> 34               dns_interface              CURLOPT_DNS_INTERFACE   string
#> 35               dns_local_ip4              CURLOPT_DNS_LOCAL_IP4   string
#> 36               dns_local_ip6              CURLOPT_DNS_LOCAL_IP6   string
#> 37                 dns_servers                CURLOPT_DNS_SERVERS   string
#> 38        dns_use_global_cache       CURLOPT_DNS_USE_GLOBAL_CACHE  integer
#> 39                   egdsocket                  CURLOPT_EGDSOCKET   string
#> 40                 errorbuffer                CURLOPT_ERRORBUFFER   string
#> 41       expect_100_timeout_ms      CURLOPT_EXPECT_100_TIMEOUT_MS  integer
#> 42                 failonerror                CURLOPT_FAILONERROR  integer
#> 43                    filetime                   CURLOPT_FILETIME  integer
#> 44                fnmatch_data               CURLOPT_FNMATCH_DATA   string
#> 45            fnmatch_function           CURLOPT_FNMATCH_FUNCTION function
#> 46              followlocation             CURLOPT_FOLLOWLOCATION  integer
#> 47                forbid_reuse               CURLOPT_FORBID_REUSE  integer
#> 48               fresh_connect              CURLOPT_FRESH_CONNECT  integer
#> 49                 ftp_account                CURLOPT_FTP_ACCOUNT   string
#> 50     ftp_alternative_to_user    CURLOPT_FTP_ALTERNATIVE_TO_USER   string
#> 51     ftp_create_missing_dirs    CURLOPT_FTP_CREATE_MISSING_DIRS  integer
#> 52              ftp_filemethod             CURLOPT_FTP_FILEMETHOD  integer
#> 53        ftp_response_timeout       CURLOPT_FTP_RESPONSE_TIMEOUT  integer
#> 54            ftp_skip_pasv_ip           CURLOPT_FTP_SKIP_PASV_IP  integer
#> 55                 ftp_ssl_ccc                CURLOPT_FTP_SSL_CCC  integer
#> 56                ftp_use_eprt               CURLOPT_FTP_USE_EPRT  integer
#> 57                ftp_use_epsv               CURLOPT_FTP_USE_EPSV  integer
#> 58                ftp_use_pret               CURLOPT_FTP_USE_PRET  integer
#> 59                     ftpport                    CURLOPT_FTPPORT   string
#> 60                  ftpsslauth                 CURLOPT_FTPSSLAUTH  integer
#> 61           gssapi_delegation          CURLOPT_GSSAPI_DELEGATION  integer
#> 62                      header                     CURLOPT_HEADER  integer
#> 63                  headerdata                 CURLOPT_HEADERDATA   string
#> 64              headerfunction             CURLOPT_HEADERFUNCTION function
#> 65                   headeropt                  CURLOPT_HEADEROPT  integer
#> 66       http_content_decoding      CURLOPT_HTTP_CONTENT_DECODING  integer
#> 67      http_transfer_decoding     CURLOPT_HTTP_TRANSFER_DECODING  integer
#> 68                http_version               CURLOPT_HTTP_VERSION  integer
#> 69              http200aliases             CURLOPT_HTTP200ALIASES   string
#> 70                    httpauth                   CURLOPT_HTTPAUTH  integer
#> 71                     httpget                    CURLOPT_HTTPGET  integer
#> 72                  httpheader                 CURLOPT_HTTPHEADER   string
#> 73                    httppost                   CURLOPT_HTTPPOST   string
#> 74             httpproxytunnel            CURLOPT_HTTPPROXYTUNNEL  integer
#> 75       ignore_content_length      CURLOPT_IGNORE_CONTENT_LENGTH  integer
#> 76                  infilesize                 CURLOPT_INFILESIZE  integer
#> 77            infilesize_large           CURLOPT_INFILESIZE_LARGE   number
#> 78                   interface                  CURLOPT_INTERFACE   string
#> 79              interleavedata             CURLOPT_INTERLEAVEDATA   string
#> 80          interleavefunction         CURLOPT_INTERLEAVEFUNCTION function
#> 81                   ioctldata                  CURLOPT_IOCTLDATA   string
#> 82               ioctlfunction              CURLOPT_IOCTLFUNCTION function
#> 83                   ipresolve                  CURLOPT_IPRESOLVE  integer
#> 84                  issuercert                 CURLOPT_ISSUERCERT   string
#> 85                   keypasswd                  CURLOPT_KEYPASSWD   string
#> 86                    krblevel                   CURLOPT_KRBLEVEL   string
#> 87                   localport                  CURLOPT_LOCALPORT  integer
#> 88              localportrange             CURLOPT_LOCALPORTRANGE  integer
#> 89               login_options              CURLOPT_LOGIN_OPTIONS   string
#> 90             low_speed_limit            CURLOPT_LOW_SPEED_LIMIT  integer
#> 91              low_speed_time             CURLOPT_LOW_SPEED_TIME  integer
#> 92                   mail_auth                  CURLOPT_MAIL_AUTH   string
#> 93                   mail_from                  CURLOPT_MAIL_FROM   string
#> 94                   mail_rcpt                  CURLOPT_MAIL_RCPT   string
#> 95        max_recv_speed_large       CURLOPT_MAX_RECV_SPEED_LARGE   number
#> 96        max_send_speed_large       CURLOPT_MAX_SEND_SPEED_LARGE   number
#> 97                 maxconnects                CURLOPT_MAXCONNECTS  integer
#> 98                 maxfilesize                CURLOPT_MAXFILESIZE  integer
#> 99           maxfilesize_large          CURLOPT_MAXFILESIZE_LARGE   number
#> 100                  maxredirs                  CURLOPT_MAXREDIRS  integer
#> 101                      netrc                      CURLOPT_NETRC  integer
#> 102                 netrc_file                 CURLOPT_NETRC_FILE   string
#> 103        new_directory_perms        CURLOPT_NEW_DIRECTORY_PERMS  integer
#> 104             new_file_perms             CURLOPT_NEW_FILE_PERMS  integer
#> 105                     nobody                     CURLOPT_NOBODY  integer
#> 106                 noprogress                 CURLOPT_NOPROGRESS  integer
#> 107                    noproxy                    CURLOPT_NOPROXY   string
#> 108                   nosignal                   CURLOPT_NOSIGNAL  integer
#> 109                 obsolete40                 CURLOPT_OBSOLETE40   string
#> 110                 obsolete72                 CURLOPT_OBSOLETE72  integer
#> 111             opensocketdata             CURLOPT_OPENSOCKETDATA   string
#> 112         opensocketfunction         CURLOPT_OPENSOCKETFUNCTION function
#> 113                   password                   CURLOPT_PASSWORD   string
#> 114                 path_as_is                 CURLOPT_PATH_AS_IS  integer
#> 115            pinnedpublickey            CURLOPT_PINNEDPUBLICKEY   string
#> 116                   pipewait                   CURLOPT_PIPEWAIT  integer
#> 117                       port                       CURLOPT_PORT  integer
#> 118                       post                       CURLOPT_POST  integer
#> 119                 postfields                 CURLOPT_POSTFIELDS   string
#> 120              postfieldsize              CURLOPT_POSTFIELDSIZE  integer
#> 121        postfieldsize_large        CURLOPT_POSTFIELDSIZE_LARGE   number
#> 122                  postquote                  CURLOPT_POSTQUOTE   string
#> 123                  postredir                  CURLOPT_POSTREDIR  integer
#> 124                   prequote                   CURLOPT_PREQUOTE   string
#> 125                    private                    CURLOPT_PRIVATE   string
#> 126               progressdata               CURLOPT_PROGRESSDATA   string
#> 127           progressfunction           CURLOPT_PROGRESSFUNCTION function
#> 128                  protocols                  CURLOPT_PROTOCOLS  integer
#> 129                      proxy                      CURLOPT_PROXY   string
#> 130         proxy_service_name         CURLOPT_PROXY_SERVICE_NAME   string
#> 131        proxy_transfer_mode        CURLOPT_PROXY_TRANSFER_MODE  integer
#> 132                  proxyauth                  CURLOPT_PROXYAUTH  integer
#> 133                proxyheader                CURLOPT_PROXYHEADER   string
#> 134              proxypassword              CURLOPT_PROXYPASSWORD   string
#> 135                  proxyport                  CURLOPT_PROXYPORT  integer
#> 136                  proxytype                  CURLOPT_PROXYTYPE  integer
#> 137              proxyusername              CURLOPT_PROXYUSERNAME   string
#> 138               proxyuserpwd               CURLOPT_PROXYUSERPWD   string
#> 139                        put                        CURLOPT_PUT  integer
#> 140                      quote                      CURLOPT_QUOTE   string
#> 141                random_file                CURLOPT_RANDOM_FILE   string
#> 142                      range                      CURLOPT_RANGE   string
#> 143                   readdata                   CURLOPT_READDATA   string
#> 144               readfunction               CURLOPT_READFUNCTION function
#> 145            redir_protocols            CURLOPT_REDIR_PROTOCOLS  integer
#> 146                    referer                    CURLOPT_REFERER   string
#> 147                    resolve                    CURLOPT_RESOLVE   string
#> 148                resume_from                CURLOPT_RESUME_FROM  integer
#> 149          resume_from_large          CURLOPT_RESUME_FROM_LARGE   number
#> 150           rtsp_client_cseq           CURLOPT_RTSP_CLIENT_CSEQ  integer
#> 151               rtsp_request               CURLOPT_RTSP_REQUEST  integer
#> 152           rtsp_server_cseq           CURLOPT_RTSP_SERVER_CSEQ  integer
#> 153            rtsp_session_id            CURLOPT_RTSP_SESSION_ID   string
#> 154            rtsp_stream_uri            CURLOPT_RTSP_STREAM_URI   string
#> 155             rtsp_transport             CURLOPT_RTSP_TRANSPORT   string
#> 156                    sasl_ir                    CURLOPT_SASL_IR  integer
#> 157                   seekdata                   CURLOPT_SEEKDATA   string
#> 158               seekfunction               CURLOPT_SEEKFUNCTION function
#> 159               service_name               CURLOPT_SERVICE_NAME   string
#> 160                      share                      CURLOPT_SHARE   string
#> 161                sockoptdata                CURLOPT_SOCKOPTDATA   string
#> 162            sockoptfunction            CURLOPT_SOCKOPTFUNCTION function
#> 163          socks5_gssapi_nec          CURLOPT_SOCKS5_GSSAPI_NEC  integer
#> 164      socks5_gssapi_service      CURLOPT_SOCKS5_GSSAPI_SERVICE   string
#> 165             ssh_auth_types             CURLOPT_SSH_AUTH_TYPES  integer
#> 166    ssh_host_public_key_md5    CURLOPT_SSH_HOST_PUBLIC_KEY_MD5   string
#> 167                ssh_keydata                CURLOPT_SSH_KEYDATA   string
#> 168            ssh_keyfunction            CURLOPT_SSH_KEYFUNCTION function
#> 169             ssh_knownhosts             CURLOPT_SSH_KNOWNHOSTS   string
#> 170        ssh_private_keyfile        CURLOPT_SSH_PRIVATE_KEYFILE   string
#> 171         ssh_public_keyfile         CURLOPT_SSH_PUBLIC_KEYFILE   string
#> 172            ssl_cipher_list            CURLOPT_SSL_CIPHER_LIST   string
#> 173               ssl_ctx_data               CURLOPT_SSL_CTX_DATA   string
#> 174           ssl_ctx_function           CURLOPT_SSL_CTX_FUNCTION function
#> 175            ssl_enable_alpn            CURLOPT_SSL_ENABLE_ALPN  integer
#> 176             ssl_enable_npn             CURLOPT_SSL_ENABLE_NPN  integer
#> 177             ssl_falsestart             CURLOPT_SSL_FALSESTART  integer
#> 178                ssl_options                CURLOPT_SSL_OPTIONS  integer
#> 179        ssl_sessionid_cache        CURLOPT_SSL_SESSIONID_CACHE  integer
#> 180             ssl_verifyhost             CURLOPT_SSL_VERIFYHOST  integer
#> 181             ssl_verifypeer             CURLOPT_SSL_VERIFYPEER  integer
#> 182           ssl_verifystatus           CURLOPT_SSL_VERIFYSTATUS  integer
#> 183                    sslcert                    CURLOPT_SSLCERT   string
#> 184                sslcerttype                CURLOPT_SSLCERTTYPE   string
#> 185                  sslengine                  CURLOPT_SSLENGINE   string
#> 186          sslengine_default          CURLOPT_SSLENGINE_DEFAULT  integer
#> 187                     sslkey                     CURLOPT_SSLKEY   string
#> 188                 sslkeytype                 CURLOPT_SSLKEYTYPE   string
#> 189                 sslversion                 CURLOPT_SSLVERSION  integer
#> 190                     stderr                     CURLOPT_STDERR   string
#> 191              tcp_keepalive              CURLOPT_TCP_KEEPALIVE  integer
#> 192               tcp_keepidle               CURLOPT_TCP_KEEPIDLE  integer
#> 193              tcp_keepintvl              CURLOPT_TCP_KEEPINTVL  integer
#> 194                tcp_nodelay                CURLOPT_TCP_NODELAY  integer
#> 195              telnetoptions              CURLOPT_TELNETOPTIONS   string
#> 196               tftp_blksize               CURLOPT_TFTP_BLKSIZE  integer
#> 197              timecondition              CURLOPT_TIMECONDITION  integer
#> 198                    timeout                    CURLOPT_TIMEOUT  integer
#> 199                 timeout_ms                 CURLOPT_TIMEOUT_MS  integer
#> 200                  timevalue                  CURLOPT_TIMEVALUE  integer
#> 201           tlsauth_password           CURLOPT_TLSAUTH_PASSWORD   string
#> 202               tlsauth_type               CURLOPT_TLSAUTH_TYPE   string
#> 203           tlsauth_username           CURLOPT_TLSAUTH_USERNAME   string
#> 204          transfer_encoding          CURLOPT_TRANSFER_ENCODING  integer
#> 205               transfertext               CURLOPT_TRANSFERTEXT  integer
#> 206           unix_socket_path           CURLOPT_UNIX_SOCKET_PATH   string
#> 207          unrestricted_auth          CURLOPT_UNRESTRICTED_AUTH  integer
#> 208                     upload                     CURLOPT_UPLOAD  integer
#> 209                        url                        CURLOPT_URL   string
#> 210                    use_ssl                    CURLOPT_USE_SSL  integer
#> 211                  useragent                  CURLOPT_USERAGENT   string
#> 212                   username                   CURLOPT_USERNAME   string
#> 213                    userpwd                    CURLOPT_USERPWD   string
#> 214                    verbose                    CURLOPT_VERBOSE  integer
#> 215              wildcardmatch              CURLOPT_WILDCARDMATCH  integer
#> 216                  writedata                  CURLOPT_WRITEDATA   string
#> 217              writefunction              CURLOPT_WRITEFUNCTION function
#> 218           xferinfofunction           CURLOPT_XFERINFOFUNCTION function
#> 219             xoauth2_bearer             CURLOPT_XOAUTH2_BEARER   string

httr_options("post")

#>                    httr                     libcurl    type
#> 26       copypostfields      CURLOPT_COPYPOSTFIELDS  string
#> 73             httppost            CURLOPT_HTTPPOST  string
#> 118                post                CURLOPT_POST integer
#> 119          postfields          CURLOPT_POSTFIELDS  string
#> 120       postfieldsize       CURLOPT_POSTFIELDSIZE integer
#> 121 postfieldsize_large CURLOPT_POSTFIELDSIZE_LARGE  number
#> 122           postquote           CURLOPT_POSTQUOTE  string
#> 123           postredir           CURLOPT_POSTREDIR integer

# Use curl_docs to read the curl documentation for each option.
# You can use either the httr or curl option name.
curl_docs("userpwd")

#> Please point your browser to the following url: 

#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html

curl_docs("CURLOPT_USERPWD")

#> Please point your browser to the following url: 

#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html