List available options.

This function lists all available options for config(). It provides both the short R name which you use with httr, and the longer Curl name, which is useful when searching the documentation. curl_doc opens a link to the libcurl documentation for an option in your browser.

Usage

httr_options(matches)

curl_docs(x)

Arguments

matches

If not missing, this restricts the output so that either the httr or curl option matches this regular expression.

x

An option name (either short or full).

Value

A data frame with three columns:

httr

The short name used in httr

libcurl

The full name used by libcurl

type

The type of R object that the option accepts

Details

RCurl and httr use slightly different names to libcurl: the initial CURLOPT_ is removed, all underscores are converted to periods and the option is given in lower case. Thus "CURLOPT_SSLENGINE_DEFAULT" becomes "sslengine.default".

Examples

httr_options()
#>                           httr                            libcurl
#> 1         abstract_unix_socket       CURLOPT_ABSTRACT_UNIX_SOCKET
#> 2              accept_encoding            CURLOPT_ACCEPT_ENCODING
#> 3             accepttimeout_ms           CURLOPT_ACCEPTTIMEOUT_MS
#> 4                address_scope              CURLOPT_ADDRESS_SCOPE
#> 5                       altsvc                     CURLOPT_ALTSVC
#> 6                  altsvc_ctrl                CURLOPT_ALTSVC_CTRL
#> 7                       append                     CURLOPT_APPEND
#> 8                  autoreferer                CURLOPT_AUTOREFERER
#> 9                    aws_sigv4                  CURLOPT_AWS_SIGV4
#> 10                  buffersize                 CURLOPT_BUFFERSIZE
#> 11                      cainfo                     CURLOPT_CAINFO
#> 12                 cainfo_blob                CURLOPT_CAINFO_BLOB
#> 13                      capath                     CURLOPT_CAPATH
#> 14                    certinfo                   CURLOPT_CERTINFO
#> 15          chunk_bgn_function         CURLOPT_CHUNK_BGN_FUNCTION
#> 16                  chunk_data                 CURLOPT_CHUNK_DATA
#> 17          chunk_end_function         CURLOPT_CHUNK_END_FUNCTION
#> 18             closesocketdata            CURLOPT_CLOSESOCKETDATA
#> 19         closesocketfunction        CURLOPT_CLOSESOCKETFUNCTION
#> 20                connect_only               CURLOPT_CONNECT_ONLY
#> 21                  connect_to                 CURLOPT_CONNECT_TO
#> 22              connecttimeout             CURLOPT_CONNECTTIMEOUT
#> 23           connecttimeout_ms          CURLOPT_CONNECTTIMEOUT_MS
#> 24  conv_from_network_function CURLOPT_CONV_FROM_NETWORK_FUNCTION
#> 25     conv_from_utf8_function    CURLOPT_CONV_FROM_UTF8_FUNCTION
#> 26    conv_to_network_function   CURLOPT_CONV_TO_NETWORK_FUNCTION
#> 27                      cookie                     CURLOPT_COOKIE
#> 28                  cookiefile                 CURLOPT_COOKIEFILE
#> 29                   cookiejar                  CURLOPT_COOKIEJAR
#> 30                  cookielist                 CURLOPT_COOKIELIST
#> 31               cookiesession              CURLOPT_COOKIESESSION
#> 32              copypostfields             CURLOPT_COPYPOSTFIELDS
#> 33                        crlf                       CURLOPT_CRLF
#> 34                     crlfile                    CURLOPT_CRLFILE
#> 35                       curlu                      CURLOPT_CURLU
#> 36               customrequest              CURLOPT_CUSTOMREQUEST
#> 37                   debugdata                  CURLOPT_DEBUGDATA
#> 38               debugfunction              CURLOPT_DEBUGFUNCTION
#> 39            default_protocol           CURLOPT_DEFAULT_PROTOCOL
#> 40                 dirlistonly                CURLOPT_DIRLISTONLY
#> 41    disallow_username_in_url   CURLOPT_DISALLOW_USERNAME_IN_URL
#> 42           dns_cache_timeout          CURLOPT_DNS_CACHE_TIMEOUT
#> 43               dns_interface              CURLOPT_DNS_INTERFACE
#> 44               dns_local_ip4              CURLOPT_DNS_LOCAL_IP4
#> 45               dns_local_ip6              CURLOPT_DNS_LOCAL_IP6
#> 46                 dns_servers                CURLOPT_DNS_SERVERS
#> 47       dns_shuffle_addresses      CURLOPT_DNS_SHUFFLE_ADDRESSES
#> 48        dns_use_global_cache       CURLOPT_DNS_USE_GLOBAL_CACHE
#> 49          doh_ssl_verifyhost         CURLOPT_DOH_SSL_VERIFYHOST
#> 50          doh_ssl_verifypeer         CURLOPT_DOH_SSL_VERIFYPEER
#> 51        doh_ssl_verifystatus       CURLOPT_DOH_SSL_VERIFYSTATUS
#> 52                     doh_url                    CURLOPT_DOH_URL
#> 53                   egdsocket                  CURLOPT_EGDSOCKET
#> 54                 errorbuffer                CURLOPT_ERRORBUFFER
#> 55       expect_100_timeout_ms      CURLOPT_EXPECT_100_TIMEOUT_MS
#> 56                 failonerror                CURLOPT_FAILONERROR
#> 57                    filetime                   CURLOPT_FILETIME
#> 58                fnmatch_data               CURLOPT_FNMATCH_DATA
#> 59            fnmatch_function           CURLOPT_FNMATCH_FUNCTION
#> 60              followlocation             CURLOPT_FOLLOWLOCATION
#> 61                forbid_reuse               CURLOPT_FORBID_REUSE
#> 62               fresh_connect              CURLOPT_FRESH_CONNECT
#> 63                 ftp_account                CURLOPT_FTP_ACCOUNT
#> 64     ftp_alternative_to_user    CURLOPT_FTP_ALTERNATIVE_TO_USER
#> 65     ftp_create_missing_dirs    CURLOPT_FTP_CREATE_MISSING_DIRS
#> 66              ftp_filemethod             CURLOPT_FTP_FILEMETHOD
#> 67        ftp_response_timeout       CURLOPT_FTP_RESPONSE_TIMEOUT
#> 68            ftp_skip_pasv_ip           CURLOPT_FTP_SKIP_PASV_IP
#> 69                 ftp_ssl_ccc                CURLOPT_FTP_SSL_CCC
#> 70                ftp_use_eprt               CURLOPT_FTP_USE_EPRT
#> 71                ftp_use_epsv               CURLOPT_FTP_USE_EPSV
#> 72                ftp_use_pret               CURLOPT_FTP_USE_PRET
#> 73                     ftpport                    CURLOPT_FTPPORT
#> 74                  ftpsslauth                 CURLOPT_FTPSSLAUTH
#> 75           gssapi_delegation          CURLOPT_GSSAPI_DELEGATION
#> 76   happy_eyeballs_timeout_ms  CURLOPT_HAPPY_EYEBALLS_TIMEOUT_MS
#> 77             haproxyprotocol            CURLOPT_HAPROXYPROTOCOL
#> 78                      header                     CURLOPT_HEADER
#> 79                  headerdata                 CURLOPT_HEADERDATA
#> 80              headerfunction             CURLOPT_HEADERFUNCTION
#> 81                   headeropt                  CURLOPT_HEADEROPT
#> 82                        hsts                       CURLOPT_HSTS
#> 83                   hsts_ctrl                  CURLOPT_HSTS_CTRL
#> 84                hstsreaddata               CURLOPT_HSTSREADDATA
#> 85            hstsreadfunction           CURLOPT_HSTSREADFUNCTION
#> 86               hstswritedata              CURLOPT_HSTSWRITEDATA
#> 87           hstswritefunction          CURLOPT_HSTSWRITEFUNCTION
#> 88              http09_allowed             CURLOPT_HTTP09_ALLOWED
#> 89              http200aliases             CURLOPT_HTTP200ALIASES
#> 90       http_content_decoding      CURLOPT_HTTP_CONTENT_DECODING
#> 91      http_transfer_decoding     CURLOPT_HTTP_TRANSFER_DECODING
#> 92                http_version               CURLOPT_HTTP_VERSION
#> 93                    httpauth                   CURLOPT_HTTPAUTH
#> 94                     httpget                    CURLOPT_HTTPGET
#> 95                  httpheader                 CURLOPT_HTTPHEADER
#> 96                    httppost                   CURLOPT_HTTPPOST
#> 97             httpproxytunnel            CURLOPT_HTTPPROXYTUNNEL
#> 98       ignore_content_length      CURLOPT_IGNORE_CONTENT_LENGTH
#> 99                  infilesize                 CURLOPT_INFILESIZE
#> 100           infilesize_large           CURLOPT_INFILESIZE_LARGE
#> 101                  interface                  CURLOPT_INTERFACE
#> 102             interleavedata             CURLOPT_INTERLEAVEDATA
#> 103         interleavefunction         CURLOPT_INTERLEAVEFUNCTION
#> 104                  ioctldata                  CURLOPT_IOCTLDATA
#> 105              ioctlfunction              CURLOPT_IOCTLFUNCTION
#> 106                  ipresolve                  CURLOPT_IPRESOLVE
#> 107                 issuercert                 CURLOPT_ISSUERCERT
#> 108            issuercert_blob            CURLOPT_ISSUERCERT_BLOB
#> 109      keep_sending_on_error      CURLOPT_KEEP_SENDING_ON_ERROR
#> 110                  keypasswd                  CURLOPT_KEYPASSWD
#> 111                   krblevel                   CURLOPT_KRBLEVEL
#> 112                  localport                  CURLOPT_LOCALPORT
#> 113             localportrange             CURLOPT_LOCALPORTRANGE
#> 114              login_options              CURLOPT_LOGIN_OPTIONS
#> 115            low_speed_limit            CURLOPT_LOW_SPEED_LIMIT
#> 116             low_speed_time             CURLOPT_LOW_SPEED_TIME
#> 117                  mail_auth                  CURLOPT_MAIL_AUTH
#> 118                  mail_from                  CURLOPT_MAIL_FROM
#> 119                  mail_rcpt                  CURLOPT_MAIL_RCPT
#> 120      mail_rcpt_alllowfails      CURLOPT_MAIL_RCPT_ALLLOWFAILS
#> 121       max_recv_speed_large       CURLOPT_MAX_RECV_SPEED_LARGE
#> 122       max_send_speed_large       CURLOPT_MAX_SEND_SPEED_LARGE
#> 123                maxage_conn                CURLOPT_MAXAGE_CONN
#> 124                maxconnects                CURLOPT_MAXCONNECTS
#> 125                maxfilesize                CURLOPT_MAXFILESIZE
#> 126          maxfilesize_large          CURLOPT_MAXFILESIZE_LARGE
#> 127           maxlifetime_conn           CURLOPT_MAXLIFETIME_CONN
#> 128                  maxredirs                  CURLOPT_MAXREDIRS
#> 129               mime_options               CURLOPT_MIME_OPTIONS
#> 130                   mimepost                   CURLOPT_MIMEPOST
#> 131                      netrc                      CURLOPT_NETRC
#> 132                 netrc_file                 CURLOPT_NETRC_FILE
#> 133        new_directory_perms        CURLOPT_NEW_DIRECTORY_PERMS
#> 134             new_file_perms             CURLOPT_NEW_FILE_PERMS
#> 135                     nobody                     CURLOPT_NOBODY
#> 136                 noprogress                 CURLOPT_NOPROGRESS
#> 137                    noproxy                    CURLOPT_NOPROXY
#> 138                   nosignal                   CURLOPT_NOSIGNAL
#> 139             opensocketdata             CURLOPT_OPENSOCKETDATA
#> 140         opensocketfunction         CURLOPT_OPENSOCKETFUNCTION
#> 141                   password                   CURLOPT_PASSWORD
#> 142                 path_as_is                 CURLOPT_PATH_AS_IS
#> 143            pinnedpublickey            CURLOPT_PINNEDPUBLICKEY
#> 144                   pipewait                   CURLOPT_PIPEWAIT
#> 145                       port                       CURLOPT_PORT
#> 146                       post                       CURLOPT_POST
#> 147                 postfields                 CURLOPT_POSTFIELDS
#> 148              postfieldsize              CURLOPT_POSTFIELDSIZE
#> 149        postfieldsize_large        CURLOPT_POSTFIELDSIZE_LARGE
#> 150                  postquote                  CURLOPT_POSTQUOTE
#> 151                  postredir                  CURLOPT_POSTREDIR
#> 152                  pre_proxy                  CURLOPT_PRE_PROXY
#> 153                   prequote                   CURLOPT_PREQUOTE
#> 154                 prereqdata                 CURLOPT_PREREQDATA
#> 155             prereqfunction             CURLOPT_PREREQFUNCTION
#> 156                    private                    CURLOPT_PRIVATE
#> 157           progressfunction           CURLOPT_PROGRESSFUNCTION
#> 158                  protocols                  CURLOPT_PROTOCOLS
#> 159                      proxy                      CURLOPT_PROXY
#> 160               proxy_cainfo               CURLOPT_PROXY_CAINFO
#> 161          proxy_cainfo_blob          CURLOPT_PROXY_CAINFO_BLOB
#> 162               proxy_capath               CURLOPT_PROXY_CAPATH
#> 163              proxy_crlfile              CURLOPT_PROXY_CRLFILE
#> 164           proxy_issuercert           CURLOPT_PROXY_ISSUERCERT
#> 165      proxy_issuercert_blob      CURLOPT_PROXY_ISSUERCERT_BLOB
#> 166            proxy_keypasswd            CURLOPT_PROXY_KEYPASSWD
#> 167      proxy_pinnedpublickey      CURLOPT_PROXY_PINNEDPUBLICKEY
#> 168         proxy_service_name         CURLOPT_PROXY_SERVICE_NAME
#> 169      proxy_ssl_cipher_list      CURLOPT_PROXY_SSL_CIPHER_LIST
#> 170          proxy_ssl_options          CURLOPT_PROXY_SSL_OPTIONS
#> 171       proxy_ssl_verifyhost       CURLOPT_PROXY_SSL_VERIFYHOST
#> 172       proxy_ssl_verifypeer       CURLOPT_PROXY_SSL_VERIFYPEER
#> 173              proxy_sslcert              CURLOPT_PROXY_SSLCERT
#> 174         proxy_sslcert_blob         CURLOPT_PROXY_SSLCERT_BLOB
#> 175          proxy_sslcerttype          CURLOPT_PROXY_SSLCERTTYPE
#> 176               proxy_sslkey               CURLOPT_PROXY_SSLKEY
#> 177          proxy_sslkey_blob          CURLOPT_PROXY_SSLKEY_BLOB
#> 178           proxy_sslkeytype           CURLOPT_PROXY_SSLKEYTYPE
#> 179           proxy_sslversion           CURLOPT_PROXY_SSLVERSION
#> 180        proxy_tls13_ciphers        CURLOPT_PROXY_TLS13_CIPHERS
#> 181     proxy_tlsauth_password     CURLOPT_PROXY_TLSAUTH_PASSWORD
#> 182         proxy_tlsauth_type         CURLOPT_PROXY_TLSAUTH_TYPE
#> 183     proxy_tlsauth_username     CURLOPT_PROXY_TLSAUTH_USERNAME
#> 184        proxy_transfer_mode        CURLOPT_PROXY_TRANSFER_MODE
#> 185                  proxyauth                  CURLOPT_PROXYAUTH
#> 186                proxyheader                CURLOPT_PROXYHEADER
#> 187              proxypassword              CURLOPT_PROXYPASSWORD
#> 188                  proxyport                  CURLOPT_PROXYPORT
#> 189                  proxytype                  CURLOPT_PROXYTYPE
#> 190              proxyusername              CURLOPT_PROXYUSERNAME
#> 191               proxyuserpwd               CURLOPT_PROXYUSERPWD
#> 192                        put                        CURLOPT_PUT
#> 193                      quote                      CURLOPT_QUOTE
#> 194                random_file                CURLOPT_RANDOM_FILE
#> 195                      range                      CURLOPT_RANGE
#> 196                   readdata                   CURLOPT_READDATA
#> 197               readfunction               CURLOPT_READFUNCTION
#> 198            redir_protocols            CURLOPT_REDIR_PROTOCOLS
#> 199                    referer                    CURLOPT_REFERER
#> 200             request_target             CURLOPT_REQUEST_TARGET
#> 201                    resolve                    CURLOPT_RESOLVE
#> 202        resolver_start_data        CURLOPT_RESOLVER_START_DATA
#> 203    resolver_start_function    CURLOPT_RESOLVER_START_FUNCTION
#> 204                resume_from                CURLOPT_RESUME_FROM
#> 205          resume_from_large          CURLOPT_RESUME_FROM_LARGE
#> 206           rtsp_client_cseq           CURLOPT_RTSP_CLIENT_CSEQ
#> 207               rtsp_request               CURLOPT_RTSP_REQUEST
#> 208           rtsp_server_cseq           CURLOPT_RTSP_SERVER_CSEQ
#> 209            rtsp_session_id            CURLOPT_RTSP_SESSION_ID
#> 210            rtsp_stream_uri            CURLOPT_RTSP_STREAM_URI
#> 211             rtsp_transport             CURLOPT_RTSP_TRANSPORT
#> 212               sasl_authzid               CURLOPT_SASL_AUTHZID
#> 213                    sasl_ir                    CURLOPT_SASL_IR
#> 214                   seekdata                   CURLOPT_SEEKDATA
#> 215               seekfunction               CURLOPT_SEEKFUNCTION
#> 216               service_name               CURLOPT_SERVICE_NAME
#> 217                      share                      CURLOPT_SHARE
#> 218                sockoptdata                CURLOPT_SOCKOPTDATA
#> 219            sockoptfunction            CURLOPT_SOCKOPTFUNCTION
#> 220                socks5_auth                CURLOPT_SOCKS5_AUTH
#> 221          socks5_gssapi_nec          CURLOPT_SOCKS5_GSSAPI_NEC
#> 222      socks5_gssapi_service      CURLOPT_SOCKS5_GSSAPI_SERVICE
#> 223             ssh_auth_types             CURLOPT_SSH_AUTH_TYPES
#> 224            ssh_compression            CURLOPT_SSH_COMPRESSION
#> 225    ssh_host_public_key_md5    CURLOPT_SSH_HOST_PUBLIC_KEY_MD5
#> 226 ssh_host_public_key_sha256 CURLOPT_SSH_HOST_PUBLIC_KEY_SHA256
#> 227                ssh_keydata                CURLOPT_SSH_KEYDATA
#> 228            ssh_keyfunction            CURLOPT_SSH_KEYFUNCTION
#> 229             ssh_knownhosts             CURLOPT_SSH_KNOWNHOSTS
#> 230        ssh_private_keyfile        CURLOPT_SSH_PRIVATE_KEYFILE
#> 231         ssh_public_keyfile         CURLOPT_SSH_PUBLIC_KEYFILE
#> 232            ssl_cipher_list            CURLOPT_SSL_CIPHER_LIST
#> 233               ssl_ctx_data               CURLOPT_SSL_CTX_DATA
#> 234           ssl_ctx_function           CURLOPT_SSL_CTX_FUNCTION
#> 235              ssl_ec_curves              CURLOPT_SSL_EC_CURVES
#> 236            ssl_enable_alpn            CURLOPT_SSL_ENABLE_ALPN
#> 237             ssl_enable_npn             CURLOPT_SSL_ENABLE_NPN
#> 238             ssl_falsestart             CURLOPT_SSL_FALSESTART
#> 239                ssl_options                CURLOPT_SSL_OPTIONS
#> 240        ssl_sessionid_cache        CURLOPT_SSL_SESSIONID_CACHE
#> 241             ssl_verifyhost             CURLOPT_SSL_VERIFYHOST
#> 242             ssl_verifypeer             CURLOPT_SSL_VERIFYPEER
#> 243           ssl_verifystatus           CURLOPT_SSL_VERIFYSTATUS
#> 244                    sslcert                    CURLOPT_SSLCERT
#> 245               sslcert_blob               CURLOPT_SSLCERT_BLOB
#> 246                sslcerttype                CURLOPT_SSLCERTTYPE
#> 247                  sslengine                  CURLOPT_SSLENGINE
#> 248          sslengine_default          CURLOPT_SSLENGINE_DEFAULT
#> 249                     sslkey                     CURLOPT_SSLKEY
#> 250                sslkey_blob                CURLOPT_SSLKEY_BLOB
#> 251                 sslkeytype                 CURLOPT_SSLKEYTYPE
#> 252                 sslversion                 CURLOPT_SSLVERSION
#> 253                     stderr                     CURLOPT_STDERR
#> 254             stream_depends             CURLOPT_STREAM_DEPENDS
#> 255           stream_depends_e           CURLOPT_STREAM_DEPENDS_E
#> 256              stream_weight              CURLOPT_STREAM_WEIGHT
#> 257   suppress_connect_headers   CURLOPT_SUPPRESS_CONNECT_HEADERS
#> 258               tcp_fastopen               CURLOPT_TCP_FASTOPEN
#> 259              tcp_keepalive              CURLOPT_TCP_KEEPALIVE
#> 260               tcp_keepidle               CURLOPT_TCP_KEEPIDLE
#> 261              tcp_keepintvl              CURLOPT_TCP_KEEPINTVL
#> 262                tcp_nodelay                CURLOPT_TCP_NODELAY
#> 263              telnetoptions              CURLOPT_TELNETOPTIONS
#> 264               tftp_blksize               CURLOPT_TFTP_BLKSIZE
#> 265            tftp_no_options            CURLOPT_TFTP_NO_OPTIONS
#> 266              timecondition              CURLOPT_TIMECONDITION
#> 267                    timeout                    CURLOPT_TIMEOUT
#> 268                 timeout_ms                 CURLOPT_TIMEOUT_MS
#> 269                  timevalue                  CURLOPT_TIMEVALUE
#> 270            timevalue_large            CURLOPT_TIMEVALUE_LARGE
#> 271              tls13_ciphers              CURLOPT_TLS13_CIPHERS
#> 272           tlsauth_password           CURLOPT_TLSAUTH_PASSWORD
#> 273               tlsauth_type               CURLOPT_TLSAUTH_TYPE
#> 274           tlsauth_username           CURLOPT_TLSAUTH_USERNAME
#> 275                trailerdata                CURLOPT_TRAILERDATA
#> 276            trailerfunction            CURLOPT_TRAILERFUNCTION
#> 277          transfer_encoding          CURLOPT_TRANSFER_ENCODING
#> 278               transfertext               CURLOPT_TRANSFERTEXT
#> 279           unix_socket_path           CURLOPT_UNIX_SOCKET_PATH
#> 280          unrestricted_auth          CURLOPT_UNRESTRICTED_AUTH
#> 281         upkeep_interval_ms         CURLOPT_UPKEEP_INTERVAL_MS
#> 282                     upload                     CURLOPT_UPLOAD
#> 283          upload_buffersize          CURLOPT_UPLOAD_BUFFERSIZE
#> 284                        url                        CURLOPT_URL
#> 285                    use_ssl                    CURLOPT_USE_SSL
#> 286                  useragent                  CURLOPT_USERAGENT
#> 287                   username                   CURLOPT_USERNAME
#> 288                    userpwd                    CURLOPT_USERPWD
#> 289                    verbose                    CURLOPT_VERBOSE
#> 290              wildcardmatch              CURLOPT_WILDCARDMATCH
#> 291                  writedata                  CURLOPT_WRITEDATA
#> 292              writefunction              CURLOPT_WRITEFUNCTION
#> 293               xferinfodata               CURLOPT_XFERINFODATA
#> 294           xferinfofunction           CURLOPT_XFERINFOFUNCTION
#> 295             xoauth2_bearer             CURLOPT_XOAUTH2_BEARER
#>         type
#> 1     string
#> 2     string
#> 3    integer
#> 4    integer
#> 5     string
#> 6    integer
#> 7    integer
#> 8    integer
#> 9     string
#> 10   integer
#> 11    string
#> 12      <NA>
#> 13    string
#> 14   integer
#> 15  function
#> 16    string
#> 17  function
#> 18    string
#> 19  function
#> 20   integer
#> 21    string
#> 22   integer
#> 23   integer
#> 24  function
#> 25  function
#> 26  function
#> 27    string
#> 28    string
#> 29    string
#> 30    string
#> 31   integer
#> 32    string
#> 33   integer
#> 34    string
#> 35    string
#> 36    string
#> 37    string
#> 38  function
#> 39    string
#> 40   integer
#> 41   integer
#> 42   integer
#> 43    string
#> 44    string
#> 45    string
#> 46    string
#> 47   integer
#> 48   integer
#> 49   integer
#> 50   integer
#> 51   integer
#> 52    string
#> 53    string
#> 54    string
#> 55   integer
#> 56   integer
#> 57   integer
#> 58    string
#> 59  function
#> 60   integer
#> 61   integer
#> 62   integer
#> 63    string
#> 64    string
#> 65   integer
#> 66   integer
#> 67   integer
#> 68   integer
#> 69   integer
#> 70   integer
#> 71   integer
#> 72   integer
#> 73    string
#> 74   integer
#> 75   integer
#> 76   integer
#> 77   integer
#> 78   integer
#> 79    string
#> 80  function
#> 81   integer
#> 82    string
#> 83   integer
#> 84    string
#> 85  function
#> 86    string
#> 87  function
#> 88   integer
#> 89    string
#> 90   integer
#> 91   integer
#> 92   integer
#> 93   integer
#> 94   integer
#> 95    string
#> 96    string
#> 97   integer
#> 98   integer
#> 99   integer
#> 100   number
#> 101   string
#> 102   string
#> 103 function
#> 104   string
#> 105 function
#> 106  integer
#> 107   string
#> 108     <NA>
#> 109  integer
#> 110   string
#> 111   string
#> 112  integer
#> 113  integer
#> 114   string
#> 115  integer
#> 116  integer
#> 117   string
#> 118   string
#> 119   string
#> 120  integer
#> 121   number
#> 122   number
#> 123  integer
#> 124  integer
#> 125  integer
#> 126   number
#> 127  integer
#> 128  integer
#> 129  integer
#> 130   string
#> 131  integer
#> 132   string
#> 133  integer
#> 134  integer
#> 135  integer
#> 136  integer
#> 137   string
#> 138  integer
#> 139   string
#> 140 function
#> 141   string
#> 142  integer
#> 143   string
#> 144  integer
#> 145  integer
#> 146  integer
#> 147   string
#> 148  integer
#> 149   number
#> 150   string
#> 151  integer
#> 152   string
#> 153   string
#> 154   string
#> 155 function
#> 156   string
#> 157 function
#> 158  integer
#> 159   string
#> 160   string
#> 161     <NA>
#> 162   string
#> 163   string
#> 164   string
#> 165     <NA>
#> 166   string
#> 167   string
#> 168   string
#> 169   string
#> 170  integer
#> 171  integer
#> 172  integer
#> 173   string
#> 174     <NA>
#> 175   string
#> 176   string
#> 177     <NA>
#> 178   string
#> 179  integer
#> 180   string
#> 181   string
#> 182   string
#> 183   string
#> 184  integer
#> 185  integer
#> 186   string
#> 187   string
#> 188  integer
#> 189  integer
#> 190   string
#> 191   string
#> 192  integer
#> 193   string
#> 194   string
#> 195   string
#> 196   string
#> 197 function
#> 198  integer
#> 199   string
#> 200   string
#> 201   string
#> 202   string
#> 203 function
#> 204  integer
#> 205   number
#> 206  integer
#> 207  integer
#> 208  integer
#> 209   string
#> 210   string
#> 211   string
#> 212   string
#> 213  integer
#> 214   string
#> 215 function
#> 216   string
#> 217   string
#> 218   string
#> 219 function
#> 220  integer
#> 221  integer
#> 222   string
#> 223  integer
#> 224  integer
#> 225   string
#> 226   string
#> 227   string
#> 228 function
#> 229   string
#> 230   string
#> 231   string
#> 232   string
#> 233   string
#> 234 function
#> 235   string
#> 236  integer
#> 237  integer
#> 238  integer
#> 239  integer
#> 240  integer
#> 241  integer
#> 242  integer
#> 243  integer
#> 244   string
#> 245     <NA>
#> 246   string
#> 247   string
#> 248  integer
#> 249   string
#> 250     <NA>
#> 251   string
#> 252  integer
#> 253   string
#> 254   string
#> 255   string
#> 256  integer
#> 257  integer
#> 258  integer
#> 259  integer
#> 260  integer
#> 261  integer
#> 262  integer
#> 263   string
#> 264  integer
#> 265  integer
#> 266  integer
#> 267  integer
#> 268  integer
#> 269  integer
#> 270   number
#> 271   string
#> 272   string
#> 273   string
#> 274   string
#> 275   string
#> 276 function
#> 277  integer
#> 278  integer
#> 279   string
#> 280  integer
#> 281  integer
#> 282  integer
#> 283  integer
#> 284   string
#> 285  integer
#> 286   string
#> 287   string
#> 288   string
#> 289  integer
#> 290  integer
#> 291   string
#> 292 function
#> 293   string
#> 294 function
#> 295   string
httr_options("post")
#>                    httr                     libcurl    type
#> 32       copypostfields      CURLOPT_COPYPOSTFIELDS  string
#> 96             httppost            CURLOPT_HTTPPOST  string
#> 130            mimepost            CURLOPT_MIMEPOST  string
#> 146                post                CURLOPT_POST integer
#> 147          postfields          CURLOPT_POSTFIELDS  string
#> 148       postfieldsize       CURLOPT_POSTFIELDSIZE integer
#> 149 postfieldsize_large CURLOPT_POSTFIELDSIZE_LARGE  number
#> 150           postquote           CURLOPT_POSTQUOTE  string
#> 151           postredir           CURLOPT_POSTREDIR integer

# Use curl_docs to read the curl documentation for each option.
# You can use either the httr or curl option name.
curl_docs("userpwd")
#> Please point your browser to the following url: 
#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html
curl_docs("CURLOPT_USERPWD")
#> Please point your browser to the following url: 
#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html