List available options.

This function lists all available options for config(). It provides both the short R name which you use with httr, and the longer Curl name, which is useful when searching the documentation. curl_doc opens a link to the libcurl documentation for an option in your browser.

Usage

httr_options(matches)

curl_docs(x)

Arguments

matches

If not missing, this restricts the output so that either the httr or curl option matches this regular expression.

x

An option name (either short or full).

Value

A data frame with three columns:

httr

The short name used in httr

libcurl

The full name used by libcurl

type

The type of R object that the option accepts

Details

RCurl and httr use slightly different names to libcurl: the initial CURLOPT_ is removed, all underscores are converted to periods and the option is given in lower case. Thus "CURLOPT_SSLENGINE_DEFAULT" becomes "sslengine.default".

Examples

httr_options()
#>                           httr                            libcurl
#> 1         abstract_unix_socket       CURLOPT_ABSTRACT_UNIX_SOCKET
#> 2              accept_encoding            CURLOPT_ACCEPT_ENCODING
#> 3             accepttimeout_ms           CURLOPT_ACCEPTTIMEOUT_MS
#> 4                address_scope              CURLOPT_ADDRESS_SCOPE
#> 5                       altsvc                     CURLOPT_ALTSVC
#> 6                  altsvc_ctrl                CURLOPT_ALTSVC_CTRL
#> 7                       append                     CURLOPT_APPEND
#> 8                  autoreferer                CURLOPT_AUTOREFERER
#> 9                   buffersize                 CURLOPT_BUFFERSIZE
#> 10                      cainfo                     CURLOPT_CAINFO
#> 11                      capath                     CURLOPT_CAPATH
#> 12                    certinfo                   CURLOPT_CERTINFO
#> 13          chunk_bgn_function         CURLOPT_CHUNK_BGN_FUNCTION
#> 14                  chunk_data                 CURLOPT_CHUNK_DATA
#> 15          chunk_end_function         CURLOPT_CHUNK_END_FUNCTION
#> 16             closesocketdata            CURLOPT_CLOSESOCKETDATA
#> 17         closesocketfunction        CURLOPT_CLOSESOCKETFUNCTION
#> 18                connect_only               CURLOPT_CONNECT_ONLY
#> 19                  connect_to                 CURLOPT_CONNECT_TO
#> 20              connecttimeout             CURLOPT_CONNECTTIMEOUT
#> 21           connecttimeout_ms          CURLOPT_CONNECTTIMEOUT_MS
#> 22  conv_from_network_function CURLOPT_CONV_FROM_NETWORK_FUNCTION
#> 23     conv_from_utf8_function    CURLOPT_CONV_FROM_UTF8_FUNCTION
#> 24    conv_to_network_function   CURLOPT_CONV_TO_NETWORK_FUNCTION
#> 25                      cookie                     CURLOPT_COOKIE
#> 26                  cookiefile                 CURLOPT_COOKIEFILE
#> 27                   cookiejar                  CURLOPT_COOKIEJAR
#> 28                  cookielist                 CURLOPT_COOKIELIST
#> 29               cookiesession              CURLOPT_COOKIESESSION
#> 30              copypostfields             CURLOPT_COPYPOSTFIELDS
#> 31                        crlf                       CURLOPT_CRLF
#> 32                     crlfile                    CURLOPT_CRLFILE
#> 33                       curlu                      CURLOPT_CURLU
#> 34               customrequest              CURLOPT_CUSTOMREQUEST
#> 35                   debugdata                  CURLOPT_DEBUGDATA
#> 36               debugfunction              CURLOPT_DEBUGFUNCTION
#> 37            default_protocol           CURLOPT_DEFAULT_PROTOCOL
#> 38                 dirlistonly                CURLOPT_DIRLISTONLY
#> 39    disallow_username_in_url   CURLOPT_DISALLOW_USERNAME_IN_URL
#> 40           dns_cache_timeout          CURLOPT_DNS_CACHE_TIMEOUT
#> 41               dns_interface              CURLOPT_DNS_INTERFACE
#> 42               dns_local_ip4              CURLOPT_DNS_LOCAL_IP4
#> 43               dns_local_ip6              CURLOPT_DNS_LOCAL_IP6
#> 44                 dns_servers                CURLOPT_DNS_SERVERS
#> 45       dns_shuffle_addresses      CURLOPT_DNS_SHUFFLE_ADDRESSES
#> 46        dns_use_global_cache       CURLOPT_DNS_USE_GLOBAL_CACHE
#> 47                     doh_url                    CURLOPT_DOH_URL
#> 48                   egdsocket                  CURLOPT_EGDSOCKET
#> 49                 errorbuffer                CURLOPT_ERRORBUFFER
#> 50       expect_100_timeout_ms      CURLOPT_EXPECT_100_TIMEOUT_MS
#> 51                 failonerror                CURLOPT_FAILONERROR
#> 52                    filetime                   CURLOPT_FILETIME
#> 53                fnmatch_data               CURLOPT_FNMATCH_DATA
#> 54            fnmatch_function           CURLOPT_FNMATCH_FUNCTION
#> 55              followlocation             CURLOPT_FOLLOWLOCATION
#> 56                forbid_reuse               CURLOPT_FORBID_REUSE
#> 57               fresh_connect              CURLOPT_FRESH_CONNECT
#> 58                 ftp_account                CURLOPT_FTP_ACCOUNT
#> 59     ftp_alternative_to_user    CURLOPT_FTP_ALTERNATIVE_TO_USER
#> 60     ftp_create_missing_dirs    CURLOPT_FTP_CREATE_MISSING_DIRS
#> 61              ftp_filemethod             CURLOPT_FTP_FILEMETHOD
#> 62        ftp_response_timeout       CURLOPT_FTP_RESPONSE_TIMEOUT
#> 63            ftp_skip_pasv_ip           CURLOPT_FTP_SKIP_PASV_IP
#> 64                 ftp_ssl_ccc                CURLOPT_FTP_SSL_CCC
#> 65                ftp_use_eprt               CURLOPT_FTP_USE_EPRT
#> 66                ftp_use_epsv               CURLOPT_FTP_USE_EPSV
#> 67                ftp_use_pret               CURLOPT_FTP_USE_PRET
#> 68                     ftpport                    CURLOPT_FTPPORT
#> 69                  ftpsslauth                 CURLOPT_FTPSSLAUTH
#> 70           gssapi_delegation          CURLOPT_GSSAPI_DELEGATION
#> 71   happy_eyeballs_timeout_ms  CURLOPT_HAPPY_EYEBALLS_TIMEOUT_MS
#> 72             haproxyprotocol            CURLOPT_HAPROXYPROTOCOL
#> 73                      header                     CURLOPT_HEADER
#> 74                  headerdata                 CURLOPT_HEADERDATA
#> 75              headerfunction             CURLOPT_HEADERFUNCTION
#> 76                   headeropt                  CURLOPT_HEADEROPT
#> 77              http09_allowed             CURLOPT_HTTP09_ALLOWED
#> 78              http200aliases             CURLOPT_HTTP200ALIASES
#> 79       http_content_decoding      CURLOPT_HTTP_CONTENT_DECODING
#> 80      http_transfer_decoding     CURLOPT_HTTP_TRANSFER_DECODING
#> 81                http_version               CURLOPT_HTTP_VERSION
#> 82                    httpauth                   CURLOPT_HTTPAUTH
#> 83                     httpget                    CURLOPT_HTTPGET
#> 84                  httpheader                 CURLOPT_HTTPHEADER
#> 85                    httppost                   CURLOPT_HTTPPOST
#> 86             httpproxytunnel            CURLOPT_HTTPPROXYTUNNEL
#> 87       ignore_content_length      CURLOPT_IGNORE_CONTENT_LENGTH
#> 88                  infilesize                 CURLOPT_INFILESIZE
#> 89            infilesize_large           CURLOPT_INFILESIZE_LARGE
#> 90                   interface                  CURLOPT_INTERFACE
#> 91              interleavedata             CURLOPT_INTERLEAVEDATA
#> 92          interleavefunction         CURLOPT_INTERLEAVEFUNCTION
#> 93                   ioctldata                  CURLOPT_IOCTLDATA
#> 94               ioctlfunction              CURLOPT_IOCTLFUNCTION
#> 95                   ipresolve                  CURLOPT_IPRESOLVE
#> 96                  issuercert                 CURLOPT_ISSUERCERT
#> 97       keep_sending_on_error      CURLOPT_KEEP_SENDING_ON_ERROR
#> 98                   keypasswd                  CURLOPT_KEYPASSWD
#> 99                    krblevel                   CURLOPT_KRBLEVEL
#> 100                  localport                  CURLOPT_LOCALPORT
#> 101             localportrange             CURLOPT_LOCALPORTRANGE
#> 102              login_options              CURLOPT_LOGIN_OPTIONS
#> 103            low_speed_limit            CURLOPT_LOW_SPEED_LIMIT
#> 104             low_speed_time             CURLOPT_LOW_SPEED_TIME
#> 105                  mail_auth                  CURLOPT_MAIL_AUTH
#> 106                  mail_from                  CURLOPT_MAIL_FROM
#> 107                  mail_rcpt                  CURLOPT_MAIL_RCPT
#> 108       max_recv_speed_large       CURLOPT_MAX_RECV_SPEED_LARGE
#> 109       max_send_speed_large       CURLOPT_MAX_SEND_SPEED_LARGE
#> 110                maxage_conn                CURLOPT_MAXAGE_CONN
#> 111                maxconnects                CURLOPT_MAXCONNECTS
#> 112                maxfilesize                CURLOPT_MAXFILESIZE
#> 113          maxfilesize_large          CURLOPT_MAXFILESIZE_LARGE
#> 114                  maxredirs                  CURLOPT_MAXREDIRS
#> 115                   mimepost                   CURLOPT_MIMEPOST
#> 116                      netrc                      CURLOPT_NETRC
#> 117                 netrc_file                 CURLOPT_NETRC_FILE
#> 118        new_directory_perms        CURLOPT_NEW_DIRECTORY_PERMS
#> 119             new_file_perms             CURLOPT_NEW_FILE_PERMS
#> 120                     nobody                     CURLOPT_NOBODY
#> 121                 noprogress                 CURLOPT_NOPROGRESS
#> 122                    noproxy                    CURLOPT_NOPROXY
#> 123                   nosignal                   CURLOPT_NOSIGNAL
#> 124                 obsolete40                 CURLOPT_OBSOLETE40
#> 125                 obsolete72                 CURLOPT_OBSOLETE72
#> 126             opensocketdata             CURLOPT_OPENSOCKETDATA
#> 127         opensocketfunction         CURLOPT_OPENSOCKETFUNCTION
#> 128                   password                   CURLOPT_PASSWORD
#> 129                 path_as_is                 CURLOPT_PATH_AS_IS
#> 130            pinnedpublickey            CURLOPT_PINNEDPUBLICKEY
#> 131                   pipewait                   CURLOPT_PIPEWAIT
#> 132                       port                       CURLOPT_PORT
#> 133                       post                       CURLOPT_POST
#> 134                 postfields                 CURLOPT_POSTFIELDS
#> 135              postfieldsize              CURLOPT_POSTFIELDSIZE
#> 136        postfieldsize_large        CURLOPT_POSTFIELDSIZE_LARGE
#> 137                  postquote                  CURLOPT_POSTQUOTE
#> 138                  postredir                  CURLOPT_POSTREDIR
#> 139                  pre_proxy                  CURLOPT_PRE_PROXY
#> 140                   prequote                   CURLOPT_PREQUOTE
#> 141                    private                    CURLOPT_PRIVATE
#> 142               progressdata               CURLOPT_PROGRESSDATA
#> 143           progressfunction           CURLOPT_PROGRESSFUNCTION
#> 144                  protocols                  CURLOPT_PROTOCOLS
#> 145                      proxy                      CURLOPT_PROXY
#> 146               proxy_cainfo               CURLOPT_PROXY_CAINFO
#> 147               proxy_capath               CURLOPT_PROXY_CAPATH
#> 148              proxy_crlfile              CURLOPT_PROXY_CRLFILE
#> 149            proxy_keypasswd            CURLOPT_PROXY_KEYPASSWD
#> 150      proxy_pinnedpublickey      CURLOPT_PROXY_PINNEDPUBLICKEY
#> 151         proxy_service_name         CURLOPT_PROXY_SERVICE_NAME
#> 152      proxy_ssl_cipher_list      CURLOPT_PROXY_SSL_CIPHER_LIST
#> 153          proxy_ssl_options          CURLOPT_PROXY_SSL_OPTIONS
#> 154       proxy_ssl_verifyhost       CURLOPT_PROXY_SSL_VERIFYHOST
#> 155       proxy_ssl_verifypeer       CURLOPT_PROXY_SSL_VERIFYPEER
#> 156              proxy_sslcert              CURLOPT_PROXY_SSLCERT
#> 157          proxy_sslcerttype          CURLOPT_PROXY_SSLCERTTYPE
#> 158               proxy_sslkey               CURLOPT_PROXY_SSLKEY
#> 159           proxy_sslkeytype           CURLOPT_PROXY_SSLKEYTYPE
#> 160           proxy_sslversion           CURLOPT_PROXY_SSLVERSION
#> 161        proxy_tls13_ciphers        CURLOPT_PROXY_TLS13_CIPHERS
#> 162     proxy_tlsauth_password     CURLOPT_PROXY_TLSAUTH_PASSWORD
#> 163         proxy_tlsauth_type         CURLOPT_PROXY_TLSAUTH_TYPE
#> 164     proxy_tlsauth_username     CURLOPT_PROXY_TLSAUTH_USERNAME
#> 165        proxy_transfer_mode        CURLOPT_PROXY_TRANSFER_MODE
#> 166                  proxyauth                  CURLOPT_PROXYAUTH
#> 167                proxyheader                CURLOPT_PROXYHEADER
#> 168              proxypassword              CURLOPT_PROXYPASSWORD
#> 169                  proxyport                  CURLOPT_PROXYPORT
#> 170                  proxytype                  CURLOPT_PROXYTYPE
#> 171              proxyusername              CURLOPT_PROXYUSERNAME
#> 172               proxyuserpwd               CURLOPT_PROXYUSERPWD
#> 173                        put                        CURLOPT_PUT
#> 174                      quote                      CURLOPT_QUOTE
#> 175                random_file                CURLOPT_RANDOM_FILE
#> 176                      range                      CURLOPT_RANGE
#> 177                   readdata                   CURLOPT_READDATA
#> 178               readfunction               CURLOPT_READFUNCTION
#> 179            redir_protocols            CURLOPT_REDIR_PROTOCOLS
#> 180                    referer                    CURLOPT_REFERER
#> 181             request_target             CURLOPT_REQUEST_TARGET
#> 182                    resolve                    CURLOPT_RESOLVE
#> 183        resolver_start_data        CURLOPT_RESOLVER_START_DATA
#> 184    resolver_start_function    CURLOPT_RESOLVER_START_FUNCTION
#> 185                resume_from                CURLOPT_RESUME_FROM
#> 186          resume_from_large          CURLOPT_RESUME_FROM_LARGE
#> 187           rtsp_client_cseq           CURLOPT_RTSP_CLIENT_CSEQ
#> 188               rtsp_request               CURLOPT_RTSP_REQUEST
#> 189           rtsp_server_cseq           CURLOPT_RTSP_SERVER_CSEQ
#> 190            rtsp_session_id            CURLOPT_RTSP_SESSION_ID
#> 191            rtsp_stream_uri            CURLOPT_RTSP_STREAM_URI
#> 192             rtsp_transport             CURLOPT_RTSP_TRANSPORT
#> 193               sasl_authzid               CURLOPT_SASL_AUTHZID
#> 194                    sasl_ir                    CURLOPT_SASL_IR
#> 195                   seekdata                   CURLOPT_SEEKDATA
#> 196               seekfunction               CURLOPT_SEEKFUNCTION
#> 197               service_name               CURLOPT_SERVICE_NAME
#> 198                      share                      CURLOPT_SHARE
#> 199                sockoptdata                CURLOPT_SOCKOPTDATA
#> 200            sockoptfunction            CURLOPT_SOCKOPTFUNCTION
#> 201                socks5_auth                CURLOPT_SOCKS5_AUTH
#> 202          socks5_gssapi_nec          CURLOPT_SOCKS5_GSSAPI_NEC
#> 203      socks5_gssapi_service      CURLOPT_SOCKS5_GSSAPI_SERVICE
#> 204             ssh_auth_types             CURLOPT_SSH_AUTH_TYPES
#> 205            ssh_compression            CURLOPT_SSH_COMPRESSION
#> 206    ssh_host_public_key_md5    CURLOPT_SSH_HOST_PUBLIC_KEY_MD5
#> 207                ssh_keydata                CURLOPT_SSH_KEYDATA
#> 208            ssh_keyfunction            CURLOPT_SSH_KEYFUNCTION
#> 209             ssh_knownhosts             CURLOPT_SSH_KNOWNHOSTS
#> 210        ssh_private_keyfile        CURLOPT_SSH_PRIVATE_KEYFILE
#> 211         ssh_public_keyfile         CURLOPT_SSH_PUBLIC_KEYFILE
#> 212            ssl_cipher_list            CURLOPT_SSL_CIPHER_LIST
#> 213               ssl_ctx_data               CURLOPT_SSL_CTX_DATA
#> 214           ssl_ctx_function           CURLOPT_SSL_CTX_FUNCTION
#> 215            ssl_enable_alpn            CURLOPT_SSL_ENABLE_ALPN
#> 216             ssl_enable_npn             CURLOPT_SSL_ENABLE_NPN
#> 217             ssl_falsestart             CURLOPT_SSL_FALSESTART
#> 218                ssl_options                CURLOPT_SSL_OPTIONS
#> 219        ssl_sessionid_cache        CURLOPT_SSL_SESSIONID_CACHE
#> 220             ssl_verifyhost             CURLOPT_SSL_VERIFYHOST
#> 221             ssl_verifypeer             CURLOPT_SSL_VERIFYPEER
#> 222           ssl_verifystatus           CURLOPT_SSL_VERIFYSTATUS
#> 223                    sslcert                    CURLOPT_SSLCERT
#> 224                sslcerttype                CURLOPT_SSLCERTTYPE
#> 225                  sslengine                  CURLOPT_SSLENGINE
#> 226          sslengine_default          CURLOPT_SSLENGINE_DEFAULT
#> 227                     sslkey                     CURLOPT_SSLKEY
#> 228                 sslkeytype                 CURLOPT_SSLKEYTYPE
#> 229                 sslversion                 CURLOPT_SSLVERSION
#> 230                     stderr                     CURLOPT_STDERR
#> 231             stream_depends             CURLOPT_STREAM_DEPENDS
#> 232           stream_depends_e           CURLOPT_STREAM_DEPENDS_E
#> 233              stream_weight              CURLOPT_STREAM_WEIGHT
#> 234   suppress_connect_headers   CURLOPT_SUPPRESS_CONNECT_HEADERS
#> 235               tcp_fastopen               CURLOPT_TCP_FASTOPEN
#> 236              tcp_keepalive              CURLOPT_TCP_KEEPALIVE
#> 237               tcp_keepidle               CURLOPT_TCP_KEEPIDLE
#> 238              tcp_keepintvl              CURLOPT_TCP_KEEPINTVL
#> 239                tcp_nodelay                CURLOPT_TCP_NODELAY
#> 240              telnetoptions              CURLOPT_TELNETOPTIONS
#> 241               tftp_blksize               CURLOPT_TFTP_BLKSIZE
#> 242            tftp_no_options            CURLOPT_TFTP_NO_OPTIONS
#> 243              timecondition              CURLOPT_TIMECONDITION
#> 244                    timeout                    CURLOPT_TIMEOUT
#> 245                 timeout_ms                 CURLOPT_TIMEOUT_MS
#> 246                  timevalue                  CURLOPT_TIMEVALUE
#> 247            timevalue_large            CURLOPT_TIMEVALUE_LARGE
#> 248              tls13_ciphers              CURLOPT_TLS13_CIPHERS
#> 249           tlsauth_password           CURLOPT_TLSAUTH_PASSWORD
#> 250               tlsauth_type               CURLOPT_TLSAUTH_TYPE
#> 251           tlsauth_username           CURLOPT_TLSAUTH_USERNAME
#> 252                trailerdata                CURLOPT_TRAILERDATA
#> 253            trailerfunction            CURLOPT_TRAILERFUNCTION
#> 254          transfer_encoding          CURLOPT_TRANSFER_ENCODING
#> 255               transfertext               CURLOPT_TRANSFERTEXT
#> 256           unix_socket_path           CURLOPT_UNIX_SOCKET_PATH
#> 257          unrestricted_auth          CURLOPT_UNRESTRICTED_AUTH
#> 258         upkeep_interval_ms         CURLOPT_UPKEEP_INTERVAL_MS
#> 259                     upload                     CURLOPT_UPLOAD
#> 260          upload_buffersize          CURLOPT_UPLOAD_BUFFERSIZE
#> 261                        url                        CURLOPT_URL
#> 262                    use_ssl                    CURLOPT_USE_SSL
#> 263                  useragent                  CURLOPT_USERAGENT
#> 264                   username                   CURLOPT_USERNAME
#> 265                    userpwd                    CURLOPT_USERPWD
#> 266                    verbose                    CURLOPT_VERBOSE
#> 267              wildcardmatch              CURLOPT_WILDCARDMATCH
#> 268                  writedata                  CURLOPT_WRITEDATA
#> 269              writefunction              CURLOPT_WRITEFUNCTION
#> 270           xferinfofunction           CURLOPT_XFERINFOFUNCTION
#> 271             xoauth2_bearer             CURLOPT_XOAUTH2_BEARER
#>         type
#> 1     string
#> 2     string
#> 3    integer
#> 4    integer
#> 5     string
#> 6    integer
#> 7    integer
#> 8    integer
#> 9    integer
#> 10    string
#> 11    string
#> 12   integer
#> 13  function
#> 14    string
#> 15  function
#> 16    string
#> 17  function
#> 18   integer
#> 19    string
#> 20   integer
#> 21   integer
#> 22  function
#> 23  function
#> 24  function
#> 25    string
#> 26    string
#> 27    string
#> 28    string
#> 29   integer
#> 30    string
#> 31   integer
#> 32    string
#> 33    string
#> 34    string
#> 35    string
#> 36  function
#> 37    string
#> 38   integer
#> 39   integer
#> 40   integer
#> 41    string
#> 42    string
#> 43    string
#> 44    string
#> 45   integer
#> 46   integer
#> 47    string
#> 48    string
#> 49    string
#> 50   integer
#> 51   integer
#> 52   integer
#> 53    string
#> 54  function
#> 55   integer
#> 56   integer
#> 57   integer
#> 58    string
#> 59    string
#> 60   integer
#> 61   integer
#> 62   integer
#> 63   integer
#> 64   integer
#> 65   integer
#> 66   integer
#> 67   integer
#> 68    string
#> 69   integer
#> 70   integer
#> 71   integer
#> 72   integer
#> 73   integer
#> 74    string
#> 75  function
#> 76   integer
#> 77   integer
#> 78    string
#> 79   integer
#> 80   integer
#> 81   integer
#> 82   integer
#> 83   integer
#> 84    string
#> 85    string
#> 86   integer
#> 87   integer
#> 88   integer
#> 89    number
#> 90    string
#> 91    string
#> 92  function
#> 93    string
#> 94  function
#> 95   integer
#> 96    string
#> 97   integer
#> 98    string
#> 99    string
#> 100  integer
#> 101  integer
#> 102   string
#> 103  integer
#> 104  integer
#> 105   string
#> 106   string
#> 107   string
#> 108   number
#> 109   number
#> 110  integer
#> 111  integer
#> 112  integer
#> 113   number
#> 114  integer
#> 115   string
#> 116  integer
#> 117   string
#> 118  integer
#> 119  integer
#> 120  integer
#> 121  integer
#> 122   string
#> 123  integer
#> 124   string
#> 125  integer
#> 126   string
#> 127 function
#> 128   string
#> 129  integer
#> 130   string
#> 131  integer
#> 132  integer
#> 133  integer
#> 134   string
#> 135  integer
#> 136   number
#> 137   string
#> 138  integer
#> 139   string
#> 140   string
#> 141   string
#> 142   string
#> 143 function
#> 144  integer
#> 145   string
#> 146   string
#> 147   string
#> 148   string
#> 149   string
#> 150   string
#> 151   string
#> 152   string
#> 153  integer
#> 154  integer
#> 155  integer
#> 156   string
#> 157   string
#> 158   string
#> 159   string
#> 160  integer
#> 161   string
#> 162   string
#> 163   string
#> 164   string
#> 165  integer
#> 166  integer
#> 167   string
#> 168   string
#> 169  integer
#> 170  integer
#> 171   string
#> 172   string
#> 173  integer
#> 174   string
#> 175   string
#> 176   string
#> 177   string
#> 178 function
#> 179  integer
#> 180   string
#> 181   string
#> 182   string
#> 183   string
#> 184 function
#> 185  integer
#> 186   number
#> 187  integer
#> 188  integer
#> 189  integer
#> 190   string
#> 191   string
#> 192   string
#> 193   string
#> 194  integer
#> 195   string
#> 196 function
#> 197   string
#> 198   string
#> 199   string
#> 200 function
#> 201  integer
#> 202  integer
#> 203   string
#> 204  integer
#> 205  integer
#> 206   string
#> 207   string
#> 208 function
#> 209   string
#> 210   string
#> 211   string
#> 212   string
#> 213   string
#> 214 function
#> 215  integer
#> 216  integer
#> 217  integer
#> 218  integer
#> 219  integer
#> 220  integer
#> 221  integer
#> 222  integer
#> 223   string
#> 224   string
#> 225   string
#> 226  integer
#> 227   string
#> 228   string
#> 229  integer
#> 230   string
#> 231   string
#> 232   string
#> 233  integer
#> 234  integer
#> 235  integer
#> 236  integer
#> 237  integer
#> 238  integer
#> 239  integer
#> 240   string
#> 241  integer
#> 242  integer
#> 243  integer
#> 244  integer
#> 245  integer
#> 246  integer
#> 247   number
#> 248   string
#> 249   string
#> 250   string
#> 251   string
#> 252   string
#> 253 function
#> 254  integer
#> 255  integer
#> 256   string
#> 257  integer
#> 258  integer
#> 259  integer
#> 260  integer
#> 261   string
#> 262  integer
#> 263   string
#> 264   string
#> 265   string
#> 266  integer
#> 267  integer
#> 268   string
#> 269 function
#> 270 function
#> 271   string
httr_options("post")
#>                    httr                     libcurl    type
#> 30       copypostfields      CURLOPT_COPYPOSTFIELDS  string
#> 85             httppost            CURLOPT_HTTPPOST  string
#> 115            mimepost            CURLOPT_MIMEPOST  string
#> 133                post                CURLOPT_POST integer
#> 134          postfields          CURLOPT_POSTFIELDS  string
#> 135       postfieldsize       CURLOPT_POSTFIELDSIZE integer
#> 136 postfieldsize_large CURLOPT_POSTFIELDSIZE_LARGE  number
#> 137           postquote           CURLOPT_POSTQUOTE  string
#> 138           postredir           CURLOPT_POSTREDIR integer

# Use curl_docs to read the curl documentation for each option.
# You can use either the httr or curl option name.
curl_docs("userpwd")
#> Please point your browser to the following url: 
#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html
curl_docs("CURLOPT_USERPWD")
#> Please point your browser to the following url: 
#> http://curl.haxx.se/libcurl/c/CURLOPT_USERPWD.html