httr_options {httr}	R Documentation

List available options.

Description

This function lists all available options for config(). It provides both the short R name which you use with httr, and the longer Curl name, which is useful when searching the documentation. curl_doc opens a link to the libcurl documentation for an option in your browser.

Usage

httr_options(matches)

curl_docs(x)

Arguments

matches
If not missing, this restricts the output so that either the httr or curl option matches this regular expression.

x
An option name (either short or full).

Details

RCurl and httr use slightly different names to libcurl: the initial CURLOPT_ is removed, all underscores are converted to periods and the option is given in lower case. Thus "CURLOPT_SSLENGINE_DEFAULT" becomes "sslengine.default".

Value

A data frame with three columns:

httr
The short name used in httr

libcurl
The full name used by libcurl

type
The type of R object that the option accepts

Examples

httr_options()
httr_options("post")

# Use curl_docs to read the curl documentation for each option.
# You can use either the httr or curl option name.
curl_docs("userpwd")
curl_docs("CURLOPT_USERPWD")

[Package httr version 1.4.7 Index]