init_oauth2.0 {httr}	R Documentation

Retrieve OAuth 2.0 access token.

Description

See demos for use.

Usage

init_oauth2.0(endpoint, app, scope = NULL, user_params = NULL,
  type = NULL, use_oob = getOption("httr_oob_default"),
  is_interactive = interactive(), use_basic_auth = FALSE)

Arguments

endpoint
An OAuth endpoint, created by oauth_endpoint

app
An OAuth consumer application, created by oauth_app

scope
a character vector of scopes to request.

user_params
Named list holding endpoint specific parameters to pass to the server when posting the request for obtaining or refreshing the access token.

type
content type used to override incorrect server response

use_oob
if FALSE, use a local webserver for the OAuth dance. Otherwise, provide a URL to the user and prompt for a validation code. Defaults to the of the "httr_oob_default" default, or TRUE if httpuv is not installed.

is_interactive
Is the current environment interactive?

use_basic_auth
if TRUE use http basic authentication to retrieve the token. Some authorization servers require this. If FALSE, the default, retrieve the token by including the app key and secret in the request body.

[Package httr version 1.2.1 Index]