Generate an oauth1.0 token.

This is the final object in the OAuth dance - it encapsulates the app, the endpoint, other parameters and the received credentials.

Usage

oauth1.0_token(
  endpoint,
  app,
  permission = NULL,
  as_header = TRUE,
  private_key = NULL,
  cache = getOption("httr_oauth_cache")
)

Arguments

endpoint

An OAuth endpoint, created by oauth_endpoint()

app

An OAuth consumer application, created by oauth_app()

permission

optional, a string of permissions to ask for.

as_header

If TRUE, the default, sends oauth in header. If FALSE, adds as parameter to url.

private_key

Optional, a key provided by openssl::read_key(). Used for signed OAuth 1.0.

cache

A logical value or a string. TRUE means to cache using the default cache file .httr-oauth, FALSE means don't cache, and NA means to guess using some sensible heuristics. A string means use the specified path as the cache file.

Value

A Token1.0 reference class (RC) object.

Details

See Token() for full details about the token object, and the caching policies used to store credentials across sessions.

See also