oauth_listener {httr}	R Documentation

Create a webserver to listen for OAuth callback.

Description

This opens a web browser pointing to request_url, and opens a webserver on port 1410 to listen to the reponse. The redirect url should either be set previously (during the OAuth authentication dance) or supplied as a parameter to the url. See oauth1.0_token() and oauth2.0_token() for examples of both techniques.

Usage

oauth_listener(request_url, is_interactive = interactive())

Arguments

request_url
the url to send the browser to

is_interactive
DEPRECATED

Details

This function should not normally be called directly by the user.

[Package httr version 1.4.7 Index]