Give verbose output.

A verbose connection provides much more information about the flow of information between the client and server.

verbose(data_out = TRUE, data_in = FALSE, info = FALSE, ssl = FALSE)

Arguments

data_out
Show data sent to the server.

data_in
Show data recieved from the server.

info
Show informational text from curl. This is mainly useful for debugging https and auth problems, so is disabled by default.

ssl
Show even data sent/recieved over SSL connections?

Prefixes

verbose() uses the following prefixes to distinguish between different components of the http messages:

* informative curl messages

-> headers sent (out)

>> data sent (out)

*> ssl data sent (out)

<- headers received (in)

<< data received (in)

<* ssl data received (in)

See also

with_verbose() makes it easier to use verbose mode even when the requests are buried inside another function call.

Other config: add_headers, authenticate, config, set_cookies, timeout, use_proxy, user_agent

Examples

GET("http://httpbin.org", verbose())

#> Response [http://httpbin.org/]
#>   Date: 2017-08-01 20:00
#>   Status: 200
#>   Content-Type: text/html; charset=utf-8
#>   Size: 12.8 kB
#> <!DOCTYPE html>
#> <html>
#> <head>
#>   <meta http-equiv='content-type' value='text/html;charset=utf8'>
#>   <meta name='generator' value='Ronn/v0.7.3 (http://github.com/rtomayko/ronn/...
#>   <title>httpbin(1): HTTP Client Testing Service</title>
#>   <style type='text/css' media='all'>
#>   /* style: man */
#>   body#manpage {margin:0}
#>   .mp {max-width:100ex;padding:0 9ex 1ex 4ex}
#> ...

GET("http://httpbin.org", verbose(info = TRUE))

#> Response [http://httpbin.org/]
#>   Date: 2017-08-01 20:00
#>   Status: 200
#>   Content-Type: text/html; charset=utf-8
#>   Size: 12.8 kB
#> <!DOCTYPE html>
#> <html>
#> <head>
#>   <meta http-equiv='content-type' value='text/html;charset=utf8'>
#>   <meta name='generator' value='Ronn/v0.7.3 (http://github.com/rtomayko/ronn/...
#>   <title>httpbin(1): HTTP Client Testing Service</title>
#>   <style type='text/css' media='all'>
#>   /* style: man */
#>   body#manpage {margin:0}
#>   .mp {max-width:100ex;padding:0 9ex 1ex 4ex}
#> ...

f <- function() {
  GET("http://httpbin.org")
}
with_verbose(f())

#> Response [http://httpbin.org/]
#>   Date: 2017-08-01 20:00
#>   Status: 200
#>   Content-Type: text/html; charset=utf-8
#>   Size: 12.8 kB
#> <!DOCTYPE html>
#> <html>
#> <head>
#>   <meta http-equiv='content-type' value='text/html;charset=utf8'>
#>   <meta name='generator' value='Ronn/v0.7.3 (http://github.com/rtomayko/ronn/...
#>   <title>httpbin(1): HTTP Client Testing Service</title>
#>   <style type='text/css' media='all'>
#>   /* style: man */
#>   body#manpage {margin:0}
#>   .mp {max-width:100ex;padding:0 9ex 1ex 4ex}
#> ...

with_verbose(f(), info = TRUE)

#> Response [http://httpbin.org/]
#>   Date: 2017-08-01 20:00
#>   Status: 200
#>   Content-Type: text/html; charset=utf-8
#>   Size: 12.8 kB
#> <!DOCTYPE html>
#> <html>
#> <head>
#>   <meta http-equiv='content-type' value='text/html;charset=utf8'>
#>   <meta name='generator' value='Ronn/v0.7.3 (http://github.com/rtomayko/ronn/...
#>   <title>httpbin(1): HTTP Client Testing Service</title>
#>   <style type='text/css' media='all'>
#>   /* style: man */
#>   body#manpage {margin:0}
#>   .mp {max-width:100ex;padding:0 9ex 1ex 4ex}
#> ...

# verbose() makes it easy to see exactly what POST requests send
POST_verbose <- function(body, ...) {
  POST("https://httpbin.org/post", body = body, verbose(), ...)
  invisible()
}
POST_verbose(list(x = "a", y = "b"))

#> 

POST_verbose(list(x = "a", y = "b"), encode = "form")

#> 

POST_verbose(FALSE)
POST_verbose(NULL)
POST_verbose("")
POST_verbose("xyz")

#>