verbose {httr}	R Documentation

Give verbose output.

Description

A verbose connection provides much more information about the flow of information between the client and server.

Usage

verbose(data_out = TRUE, data_in = FALSE, info = FALSE, ssl = FALSE)

Arguments

data_out
Show data sent to the server.

data_in
Show data recieved from the server.

info
Show informational text from curl. This is mainly useful for debugging https and auth problems, so is disabled by default.

ssl
Show even data sent/recieved over SSL connections?

Prefixes

verbose() uses the following prefixes to distinguish between different components of the http messages:

* informative curl messages

⁠->⁠ headers sent (out)

⁠>>⁠ data sent (out)

⁠*>⁠ ssl data sent (out)

⁠<-⁠ headers received (in)

⁠<<⁠ data received (in)

⁠<*⁠ ssl data received (in)

See Also

with_verbose() makes it easier to use verbose mode even when the requests are buried inside another function call.

Other config: add_headers(), authenticate(), config(), set_cookies(), timeout(), use_proxy(), user_agent()

Examples

## Not run: 
GET("http://httpbin.org", verbose())
GET("http://httpbin.org", verbose(info = TRUE))

f <- function() {
  GET("http://httpbin.org")
}
with_verbose(f())
with_verbose(f(), info = TRUE)

# verbose() makes it easy to see exactly what POST requests send
POST_verbose <- function(body, ...) {
  POST("https://httpbin.org/post", body = body, verbose(), ...)
  invisible()
}
POST_verbose(list(x = "a", y = "b"))
POST_verbose(list(x = "a", y = "b"), encode = "form")
POST_verbose(FALSE)
POST_verbose(NULL)
POST_verbose("")
POST_verbose("xyz")

## End(Not run)

[Package httr version 1.4.7 Index]